Base

PowerModelsDistribution.ref — Function

helper function to access the AbstractUnbalancedPowerModel structs' ref, returns (id,data) pairs

PowerModelsDistribution.var — Function

helper function to access the AbstractUnbalancedPowerModel structs' var, returns JuMP VariableRef

PowerModelsDistribution.ids — Function

helper function to access the ids of AbstractUnbalancedPowerModel structs' ref, returns ints

PowerModelsDistribution.con — Function

helper function to access the AbstractUnbalancedPowerModel structs' con, returns JuMP Constraint

PowerModelsDistribution.sol — Function

helper function to access the AbstractUnbalancedPowerModel structs' sol, returns Dict

PowerModelsDistribution.nws — Function

helper function to access multinetwork data from AbstractUnbalancedPowerModel structs, returns (id,data) pairs

PowerModelsDistribution.nw_ids — Function

helper function to access the ids of multinetworks from AbstractUnbalancedPowerModel structs, returns ints

Helper functions

PowerModelsDistribution.@smart_constraint — Macro

@smart_constraint model::JuMP.Model vars::Vector expr::JuMP.Expression

Detection of whether a constraint should be NL or not"

PowerModelsDistribution.set_lower_bound — Function

function set_lower_bound(
	x::JuMP.VariableRef,
	bound::Real
)

Local wrapper method for JuMP.setlowerbound, which skips NaN and infinite (-Inf only)

function set_lower_bound(
	xs::Vector{JuMP.VariableRef},
	bound::Real
)

Local wrapper method for JuMP.setlowerbound, which skips NaN and infinite (-Inf only). Note that with this signature, the bound is applied to every variable in the vector.

PowerModelsDistribution.set_upper_bound — Function

function set_upper_bound(
	x::JuMP.VariableRef,
	bound
)

Local wrapper method for JuMP.setupperbound, which skips NaN and infinite (+Inf only)

function set_upper_bound(
	xs::Vector{JuMP.VariableRef},
	bound::Real
)

Local wrapper method for JuMP.setupperbound, which skips NaN and infinite (+Inf only). Note that with this signature, the bound is applied to every variable in the vector.

PowerModelsDistribution.comp_start_value — Function

comp_start_value(
  comp::Dict, 
  keys::Vector{String}, 
  conductor::Int, 
  default::Any
)

Searches for start value for a variable key in order from a list of keys of a component comp for conductor conductor, and if one does not exist, uses default

function comp_start_value(
	comp::Dict{String,<:Any},
	key::String,
	conductor::Int,
	default::Any
)

Searches for start value for a variable key of a component comp for conductor conductor, and if one does not exist, uses default

function comp_start_value(
	comp::Dict{String,<:Any},
	keys::String,
	default::Any=0.0
)

Searches for start value for a variable key in order from a list of keys of a component comp, and if one does not exist, uses default. This is the conductor-agnostic version of comp_start_value.

function comp_start_value(
  comp::Dict, 
  key::String, 
  default::Any=0.0
)

Searches for start value for a variable key of a component comp, and if one does not exist, uses default. This is the conductor-agnostic version of comp_start_value.

Ref Creation Functions

PowerModelsDistribution.ref_add_connected_components! — Method

ref_add_connected_components!(ref::Dict{Symbol,<:Any}, data::Dict{String,<:Any})

Ref-extension for opb problem type to add connected components ref

PowerModelsDistribution.ref_add_core! — Method

ref_add_core!(ref::Dict{Symbol,Any})

Returns a dict that stores commonly used pre-computed data from of the data dictionary, primarily for converting data-types, filtering out deactivated components, and storing system-wide values that need to be computed globally. Some of the common keys include:

:off_angmin and :off_angmax (see calc_theta_delta_bounds(data)),

:bus – the set {(i, bus) in ref[:bus] : bus["bus_type"] != 4},

:gen – the set {(i, gen) in ref[:gen] : gen["gen_status"] == 1 && gen["gen_bus"] in keys(ref[:bus])},

:branch – the set of branches that are active in the network (based on the component status values),

:arcs_branch_from – the set [(i,b["f_bus"],b["t_bus"]) for (i,b) in ref[:branch]],

:arcs_branch_to – the set [(i,b["t_bus"],b["f_bus"]) for (i,b) in ref[:branch]],

:arcs_branch – the set of arcs from both arcs_from and arcs_to,

:arcs_switch_from – the set [(i,b["f_bus"],b["t_bus"]) for (i,b) in ref[:switch]],

:arcs_switch_to – the set [(i,b["t_bus"],b["f_bus"]) for (i,b) in ref[:switch]],

:arcs_switch – the set of arcs from both arcs_switch_from and arcs_switch_to,

:arcs_transformer_from – the set [(i,b["f_bus"],b["t_bus"]) for (i,b) in ref[:transformer]],

:arcs_transformer_to – the set [(i,b["t_bus"],b["f_bus"]) for (i,b) in ref[:transformer]],

:arcs_transformer – the set of arcs from both arcs_transformer_from and arcs_transformer_to,

:bus_arcs_branch – the mapping Dict(i => [(l,i,j) for (l,i,j) in ref[:arcs_branch]]),

:bus_arcs_transformer – the mapping Dict(i => [(l,i,j) for (l,i,j) in ref[:arcs_transformer]]),

:bus_arcs_switch – the mapping Dict(i => [(l,i,j) for (l,i,j) in ref[:arcs_switch]]),

:buspairs – (see buspair_parameters(ref[:arcs_branch_from], ref[:branch], ref[:bus])),

:bus_gens – the mapping Dict(i => [gen["gen_bus"] for (i,gen) in ref[:gen]]).

:bus_loads – the mapping Dict(i => [load["load_bus"] for (i,load) in ref[:load]]).

:bus_shunts – the mapping Dict(i => [shunt["shunt_bus"] for (i,shunt) in ref[:shunt]]).

PowerModelsDistribution.ref_calc_storage_injection_bounds — Method

ref_calc_storage_injection_bounds(storage, buses)

Computes storage bounds

InfrastructureModels Extensions

InfrastructureModels.solution_preprocessor — Function

Definition of the default solution preprocessor for PowerModelsDistribution

InfrastructureModels.build_solution_values — Function

custom build_solution_values for multiconductor (vector) variables

custom build_solution_values for multiconductor (vector) nonlinear expressions

custom build_solution_values for multiconductor (vector) generic affine expressions

custom build_solution_values for multiconductor (vector) constants

custom build_solution_values for generic dense axis arrays

custom build_solution_values for multiconductor (vector) constants