Enums

PowerModelsDistribution.PowerModelsDistributionEnums — Type

Collection of the built-in Enums for PowerModelsDistribution

PowerModelsDistribution.DataModel — Type

DataModel

An Enum to descibe the current data model contained in the structure

PowerModelsDistribution.ENGINEERING — Constant

Model type for models that are in the PowerModelsDistribution engineering representation

PowerModelsDistribution.MATHEMATICAL — Constant

Model type for models that are in the mathematical representation

PowerModelsDistribution.DSS — Constant

Model type for raw dss imports

PowerModelsDistribution.MATPOWER — Constant

Model type for models imported via parse_file from PowerModels

PowerModelsDistribution.LoadModel — Type

LoadModel

An Enum to describe the type of load, e.g., constant power, constant current, etc.

Information about load models can be found under Load Models

PowerModelsDistribution.POWER — Constant

Constant Power load model

PowerModelsDistribution.CURRENT — Constant

Constant Current load model

PowerModelsDistribution.IMPEDANCE — Constant

Constant Impedance load model

PowerModelsDistribution.EXPONENTIAL — Constant

Exponential load model

PowerModelsDistribution.ZIP — Constant

ZIP load model

PowerModelsDistribution.ShuntModel — Type

ShuntModel

An Enum to describe the type of shunt, e.g., generic, capcitor or reactor type.

PowerModelsDistribution.GENERIC — Constant

Generic shunt model, usually indicates user-created

PowerModelsDistribution.CAPACITOR — Constant

Capacitor shunt model, usually indicates parsed from capacitor dss object

PowerModelsDistribution.REACTOR — Constant

Reactor shunt model, usually indicates parsed from reactor dss object

PowerModelsDistribution.SwitchState — Type

SwitchState

An Enum to descibe whether a switch is open or closed

PowerModelsDistribution.OPEN — Constant

Switch state is open

PowerModelsDistribution.CLOSED — Constant

Switch state is closed

PowerModelsDistribution.ControlMode — Type

ControlMode

An Enum to descibe the current control mode of the generation object

PowerModelsDistribution.FREQUENCYDROOP — Constant

Generation resource is in frequency droop mode (following)

PowerModelsDistribution.ISOCHRONOUS — Constant

Generation resource is in isochronous mode (forming)

PowerModelsDistribution.ConnConfig — Type

ConnConfig

An Enum to describe the connection configuration, e.g., wye or delta

PowerModelsDistribution.WYE — Constant

Wye connection configuration

PowerModelsDistribution.DELTA — Constant

Delta connection configuration

PowerModelsDistribution.Dispatchable — Type

Dispatchable

An Enum to describe whether an object is dispatchable, e.g., can a switch state be controled, or can a load or shunt be shed individually.

PowerModelsDistribution.NO — Constant

The object is not dispatchable / controllable

PowerModelsDistribution.YES — Constant

The object is dispatchable / controllable

PowerModelsDistribution.Status — Type

Status

An Enum to describe whether an object is enabled or disabled

PowerModelsDistribution.DISABLED — Constant

The object is disabled

PowerModelsDistribution.ENABLED — Constant

The object is enabled