Formulations

Abstract Models

PowerModelsDistribution.AbstractUnbalancedPowerModel — Type

root of the power formulation type hierarchy

PowerModelsDistribution.AbstractUnbalancedActivePowerModel — Type

active power only models

PowerModelsDistribution.AbstractUnbalancedConicModel — Type

variants that target conic solvers

PowerModelsDistribution.AbstractUBFModel — Type

for branch flow models

PowerModelsDistribution.AbstractUBFAModel — Type

for variants of branch flow models that target LP solvers

PowerModelsDistribution.AbstractUBFQPModel — Type

for variants of branch flow models that target QP or NLP solvers

PowerModelsDistribution.AbstractUBFConicModel — Type

for variants of branch flow models that target conic solvers

PowerModelsDistribution.AbstractUnbalancedACPModel — Type

Abstract Power-Voltage (Polar) formulation

PowerModelsDistribution.AbstractUnbalancedACRModel — Type

Abstract Power-Voltage (Rectangular) formulation

PowerModelsDistribution.AbstractUnbalancedIVRModel — Type

Abstract Current-Voltage (Rectangular) formulation

PowerModelsDistribution.AbstractUnbalancedDCPModel — Type

PowerModelsDistribution.AbstractUnbalancedNFAModel — Type

PowerModelsDistribution.AbstractNLPUBFModel — Type

Base Abstract NLP Unbalanced Branch Flow Model

PowerModelsDistribution.AbstractConicUBFModel — Type

Base Abstract Conic Unbalanced Branch Flow Model

PowerModelsDistribution.SDPUBFModel — Type

SDP BFM per Gan and Low 2014, PSCC

PowerModelsDistribution.SDPUBFKCLMXModel — Type

SDP BFM with KCL as matrix equation, Geth 2020 (under review)

PowerModelsDistribution.SOCNLPUBFModel — Type

SOC relaxation of SDPUBFModel per Kim, Kojima, & Yamashita 2003, cast as an QCP

PowerModelsDistribution.SOCConicUBFModel — Type

SOC relaxation of SDPUBFModel per Kim, Kojima, & Yamashita 2003, cast as a SOC

PowerModelsDistribution.AbstractLPUBFModel — Type

Abstract form for linear unbalanced power flow models

PowerModelsDistribution.LPUBFDiagModel — Type

LinDist3Flow per Arnold et al. (2016), using vector variables for power, voltage and current

D. B. Arnold, M. Sankur, R. Dobbe, K. Brady, D. S. Callaway and A. Von Meier, "Optimal dispatch of reactive power for voltage regulation and balancing in unbalanced distribution systems," 2016 IEEE Power and Energy Society General Meeting (PESGM), Boston, MA, 2016, pp. 1-5, doi: 10.1109/PESGM.2016.7741261.

PowerModelsDistribution.LinDist3FlowModel — Type

More popular name for the LPUBFDiagModel

Abstract Union Models

PowerModelsDistribution.AbstractUBFModels — Type

Collection of Unbalanced Branch Flow Models

PowerModelsDistribution.KCLMXModels — Type

Collection of Semidefinite Models

PowerModelsDistribution.SOCUBFModels — Type

Collection of Second Order Cone Models

PowerModelsDistribution.AbstractUnbalancedWModels — Type

Collection of AbstractUnbalancedPowerModels that include W relaxations

PowerModelsDistribution.AbstractUnbalancedWConvexModels — Type

Collection of convex AbstractUnbalancedPowerModels that include W relaxations

PowerModelsDistribution.AbstractUnbalancedAPLossLessModels — Type

Collection of AbstractUnbalancedPowerModels that are Active Power only and Lossless

PowerModelsDistribution.AbstractUnbalancedPolarModels — Type

Collection of AbstractUnbalancedPowerModels that have a Polar representation

Power Models

PowerModelsDistribution.ACPUPowerModel — Type

AC power flow Model with polar bus voltage variables. The seminal reference of AC OPF:

@article{carpentier1962contribution,
  title={Contribution to the economic dispatch problem},
  author={Carpentier, J},
  journal={Bulletin de la Societe Francoise des Electriciens},
  volume={3},
  number={8},
  pages={431--447},
  year={1962}
}

History and discussion:

@techreport{Cain2012,
  author = {Cain, Mary B and {O' Neill}, Richard P and Castillo, Anya},
  title = {{History of optimal power flow and Models}},
  year = {2012}
  pages = {1--36},
  url = {https://www.ferc.gov/industries/electric/indus-act/market-planning/opf-papers/acopf-1-history-Model-testing.pdf}
}

PowerModelsDistribution.ACRUPowerModel — Type

AC power flow Model with rectangular bus voltage variables.

@techreport{Cain2012,
  author = {Cain, Mary B and {O' Neill}, Richard P and Castillo, Anya},
  pages = {1--36},
  title = {{History of optimal power flow and Models}},
  url = {https://www.ferc.gov/industries/electric/indus-act/market-planning/opf-papers/acopf-1-history-Model-testing.pdf}
  year = {2012}
}

PowerModelsDistribution.IVRUPowerModel — Type

Current voltage formulation of AC OPF. The formulation uses rectangular coordinates for both current and voltage. Note that, even though Kirchhoff's circuit laws are linear in current and voltage, this formulation is nonconvex due to constants power loads/generators and apparent power limits.

@techreport{ONeill2012,
    author = {{O' Neill}, Richard P and Castillo, Anya and Cain, Mary B},
    pages = {1--18},
    title = {{The IV formulation and linear approximations of the ac optimal power flow problem}},
    year = {2012}
}

Applicable to problem formulations with _iv in the name.

PowerModelsDistribution.DCPUPowerModel — Type

Linearized 'DC' power flow Model with polar voltage variables. This model is a basic linear active-power-only approximation, which uses branch susceptance values br_b = -br_x / (br_x^2 + br_x^2) for determining the network phase angles. Furthermore, transformer parameters such as tap ratios and phase shifts are not considered as part of this model. It is important to note that it is also common for active-power-only approximations to use 1/br_x for determining the network phase angles, instead of the br_b value that is used here. Small discrepancies in solutions should be expected when comparing active-power-only approximations across multiple tools.

@ARTICLE{4956966,
  author={B. Stott and J. Jardim and O. Alsac},
  journal={IEEE Transactions on Power Systems},
  title={DC Power Flow Revisited},
  year={2009},
  month={Aug},
  volume={24},
  number={3},
  pages={1290-1300},
  doi={10.1109/TPWRS.2009.2021235},
  ISSN={0885-8950}
}

PowerModelsDistribution.NFAUPowerModel — Type

The an active power only network flow approximation, also known as the transportation model.

PowerModelsDistribution.LPUBFDiagPowerModel — Type

default LP unbalanced DistFlow constructor

PowerModelsDistribution.LinDist3FlowPowerModel — Type

More popular name for the LPUBFDiagPowerModel

PowerModelsDistribution.SDPUBFPowerModel — Type

default SDP unbalanced DistFlow constructor

PowerModelsDistribution.SDPUBFKCLMXPowerModel — Type

default SDP unbalanced DistFlow with matrix KCL constructor

PowerModelsDistribution.SOCNLPUBFPowerModel — Type

default SOC unbalanced DistFlow constructor

PowerModelsDistribution.SOCConicUBFPowerModel — Type

default SOC unbalanced DistFlow constructor

Mutable Stuct creator

PowerModelsDistribution.@pmd_fields — Macro

a macro for adding the base PowerModels fields to a type definition