Internal Functions

PowerModelsDistribution._1to1_maps — Constant

items that are mapped one-to-one from engineering to math models

PowerModelsDistribution._array_delimiters — Constant

different acceptable delimiters for arrays

PowerModelsDistribution._conductor_matrix — Constant

field names that should become multi-conductor matrix not arrays

PowerModelsDistribution._conductorless — Constant

field names that should not be multi-conductor values

PowerModelsDistribution._convert_to_meters — Constant

conversion factors for units to meters

PowerModelsDistribution._double_operators — Constant

two number operators for reverse polish notation

PowerModelsDistribution._dss2pmd_capcontrol_type — Constant

dss to pmd capcontrol type

PowerModelsDistribution._dss2pmd_load_model — Constant

dss to pmd load model

PowerModelsDistribution._dss_control_objects — Constant

all objects that define controls

PowerModelsDistribution._dss_data_objects — Constant

all data holding objects

PowerModelsDistribution._dss_edge_objects — Constant

all edge types that can help define buses

PowerModelsDistribution._dss_monitor_objects — Constant

all objects that provide montoring

PowerModelsDistribution._dss_node_objects — Constant

all node types that can help define buses

PowerModelsDistribution._dss_parameter_data_types — Constant

Returns a Dict{String,Type} for the desired component comp, giving all of the expected data types

PowerModelsDistribution._dss_supported_components — Constant

components currently supported for automatic data type parsing

PowerModelsDistribution._dtype_regex — Constant

Regexes for determining data types

PowerModelsDistribution._eng_edge_elements — Constant

list of edge type elements in the engineering model

PowerModelsDistribution._eng_model_checks — Constant

data check functions for the engineering data model

PowerModelsDistribution._eng_model_dtypes — Constant

Data types of accepted fields in the engineering data model

PowerModelsDistribution._eng_model_req_fields — Constant

required fields in the engineering data model

PowerModelsDistribution._eng_node_elements — Constant

list of nodal type elements in the engineering model

PowerModelsDistribution._like_exclusions — Constant

properties that should be excluded from being overwritten during the application of like

PowerModelsDistribution._loss_model_objects — Constant

collect of components and their properties that define loss models when transforming to the MATHEMATICAL model

PowerModelsDistribution._math_dispatchable_elements — Constant

list of math asset types that are dispatchable

PowerModelsDistribution._math_edge_elements — Constant

list of edge type elements in the engineering model

PowerModelsDistribution._math_node_elements — Constant

list of nodal type elements in the engineering model

PowerModelsDistribution._pmd_math_global_keys — Constant

list of multinetwork keys that belong at the root level

PowerModelsDistribution._single_operators — Constant

single number operators in reverse polish notation

InfrastructureModels.build_solution_values — Method

custom build_solution_values for generic dense axis arrays

InfrastructureModels.build_solution_values — Method

custom build_solution_values for multiconductor (vector) generic affine expressions

InfrastructureModels.build_solution_values — Method

custom build_solution_values for multiconductor (vector) nonlinear expressions

InfrastructureModels.build_solution_values — Method

custom build_solution_values for multiconductor (vector) variables

InfrastructureModels.build_solution_values — Method

custom build_solution_values for multiconductor (vector) constants

InfrastructureModels.build_solution_values — Method

custom build_solution_values for multiconductor (vector) constants

InfrastructureModels.solution_preprocessor — Method

Definition of the default solution preprocessor for PowerModelsDistribution

PowerModelsDistribution._add_component! — Method

Adds a component of type obj_type_name with properties given by object to the existing data_dss structure. If a component of the same type has already been added to data_dss, the new component is appeneded to the existing array of components of that type, otherwise a new array is created.

PowerModelsDistribution._add_eng_obj! — Method

add engineering data object to engineering data model

PowerModelsDistribution._add_gen_cost_model! — Method

convert cost model names

PowerModelsDistribution._add_implicit_absolute_bounds! — Method

When a terminal is grounded, any pairwise bounds it occurs in imply an absolute bound for the other terminal in the pair. This method converts such pairwise bounds to absolute ones.

PowerModelsDistribution._add_property — Method

Adds a property to an existing component properties dictionary object given the key and value of the property. If a property of the same name already exists inside object, the original value is converted to an array, and the new value is appended to the end.

PowerModelsDistribution._add_unused_kwargs! — Method

adds kwargs that were specified but unused by the required defaults to the component

PowerModelsDistribution._adjust_line_limits! — Method

_adjust_line_limits!(data_eng::Dict{String,<:Any}, mult::Float64)

Multiplies limits (sm_ub and/or cm_ub) on line objects (line, linecode, switch) by a multiplier mult

PowerModelsDistribution._adjust_small_line_admittances! — Method

_adjust_small_line_admittances!(data_eng::Dict{String,<:Any}; min_admittance_val::Float64=1e-2, replace_admittance_val::Float64=0.0)

Replaces admittances (gfr, gto, bfr, bto) on lines, linecodes, and switches lower than min_admittance_val with replace_admittance_val.

PowerModelsDistribution._adjust_small_line_impedances! — Method

_adjust_small_line_impedances!(data_eng::Dict{String,<:Any}; min_impedance_val::Float64=1e-2, replace_impedance_val::Float64=0.0)

Replaces impedances (rs, xs) on lines, linecodes, and switches lower than min_impedance_val with replace_impedance_val.

PowerModelsDistribution._adjust_small_line_lengths! — Method

_adjust_small_line_lengths!(data_eng::Dict{String,<:Any}; min_length_val::Float64=25.0, replace_length_val::Float64=0.0)

Replaces length on lines, switches lower than min_length_val with replace_length_val.

PowerModelsDistribution._adjust_transformer_limits! — Method

_adjust_transformer_limits!(data_eng::Dict{String,<:Any}, mult::Float64)

Multiplies limits (sm_ub and/or cm_ub) on transformer objects by a multiplier mult

PowerModelsDistribution._admittance_conversion — Method

converts admittance by multiplying by 2πωl

PowerModelsDistribution._apply_filter! — Method

Filters out values of a vector or matrix for certain properties (transformer winding variant)

PowerModelsDistribution._apply_filter! — Method

Filters out values of a vector or matrix for certain properties

PowerModelsDistribution._apply_func_vals — Method

helper to apply function values

PowerModelsDistribution._apply_kron_reduction! — Method

_apply_kron_reduction!(data_eng::Dict{String,<:Any}; kr_phases::Union{Vector{Int},Vector{String}}=[1,2,3], kr_neutral::Union{Int,String}=4)

Applies a Kron Reduction to the network, reducing out the kr_neutral, leaving only the kr_phases

PowerModelsDistribution._apply_like! — Method

applies like to component

PowerModelsDistribution._apply_linecode! — Method

applies a linecode to a line in preparation for converting to mathematical model

PowerModelsDistribution._apply_ordered_properties — Method

apply properties in the order that they are given

PowerModelsDistribution._apply_phase_projection! — Method

_apply_phase_projection!(data_eng::Dict{String,<:Any})

pad matrices and vectors to max number of conductors

PowerModelsDistribution._apply_phase_projection_delta! — Method

_apply_phase_projection_delta!(data_eng::Dict{String,<:Any})

phase projection for components where unprojected states are not yet supported (delta configurations).

See apply_phase_projection!

PowerModelsDistribution._apply_voltage_angle_difference_bounds! — Function

_apply_voltage_angle_difference_bounds!(eng::Dict{String,<:Any}, vad::Real=5.0)

Applies voltage angle difference bound given by vad::Real in degrees (i.e., the allowed drift of angle from one end of a line to another) to all lines. By default, vad=5.0.

PowerModelsDistribution._apply_voltage_bounds! — Method

_apply_voltage_bounds!(data_eng::Dict{String,<:Any}; vm_lb::Union{Real,Missing}=0.9, vm_ub::Union{Real,Missing}=1.1)

add voltage bounds to all buses based on per-unit upper (vm_ub) and lower (vm_lb), scaled by the bus's voltage based

PowerModelsDistribution._apply_xfmrcode! — Method

applies a xfmrcode to a transformer in preparation for converting to mathematical model

PowerModelsDistribution._assign_property! — Method

Assigns a property with name property_name and value property_value to the component of type obj_type named obj_name in data_dss.

PowerModelsDistribution._bank_transformers! — Method

Combines transformers with 'bank' keyword into a single transformer

PowerModelsDistribution._barrel_roll — Method

shifts a vector by shift spots to the left

PowerModelsDistribution._biggest_der — Method

finds the largest active generation asset (gen, storage) in an island

PowerModelsDistribution._biggest_generator — Method

find the largest active generator in a collection of generators

PowerModelsDistribution._build_bus_shunt_matrices — Method

helper function to build bus shunt matrices for power balance constraints

PowerModelsDistribution._build_loss_model! — Method

loss model builder for transformer decomposition

PowerModelsDistribution._build_mc_mn_opb — Method

_build_mc_mn_opb(pm::AbstractUnbalancedPowerModel)

Constructor for Optimal Power Balance

PowerModelsDistribution._build_mc_osw — Method

constructor for branch flow osw

PowerModelsDistribution._build_mc_osw — Method

constructor for OSW in current-voltage variable space

PowerModelsDistribution._build_mc_osw — Method

Constructor for Optimal Switching

PowerModelsDistribution._build_mc_osw_mi — Method

_build_mc_osw_mi(pm::AbstractUBFModels)

constructor for mixed-integer branch flow osw

PowerModelsDistribution._build_mc_osw_mi — Method

Constructor for Optimal Switching

PowerModelsDistribution._build_time_series_reference! — Method

helper function to properly reference time series variables from opendss

PowerModelsDistribution._calc_branch_current_max — Method

Returns a total (shunt+series) current magnitude bound for the from and to side of a branch. The total power rating also implies a current bound through the lower bound on the voltage magnitude of the connected buses.

PowerModelsDistribution._calc_branch_current_max_frto — Method

Returns a total (shunt+series) current magnitude bound for the from and to side of a branch. The total power rating also implies a current bound through the lower bound on the voltage magnitude of the connected buses.

PowerModelsDistribution._calc_branch_power_max — Method

Returns a total (shunt+series) power magnitude bound for the from and to side of a branch. The total current rating also implies a current bound through the upper bound on the voltage magnitude of the connected buses.

PowerModelsDistribution._calc_branch_power_max_frto — Method

Returns a total (shunt+series) power magnitude bound for the from and to side of a branch. The total current rating also implies a current bound through the upper bound on the voltage magnitude of the connected buses.

PowerModelsDistribution._calc_branch_series_current_max — Method

Returns a valid series current magnitude bound for a branch.

PowerModelsDistribution._calc_bus_vm_ll_bounds — Method

Returns bounds in line-to-line bounds on the voltage magnitude. If these are not part of the problem specification, then a valid upper bound is implied by the line-to-neutral bounds, but a lower bound (greater than zero) is not. Therefore, a default lower bound is then used, specified by the keyword argument vdmin_eps. The returned bounds are for the pairs 1->2, 2->3, 3->1

PowerModelsDistribution._calc_comp_lines — Method

_calc_comp_lines(component::Dict{String,<:Any})

compute lines in m and b from from pwl cost models

PowerModelsDistribution._calc_connected_components_eng — Method

computes the connected components of the network graph returns a set of sets of bus ids, each set is a connected component

PowerModelsDistribution._calc_connected_components_math — Method

computes the connected components of the network graph returns a set of sets of bus ids, each set is a connected component

PowerModelsDistribution._calc_gen_current_max — Method

Returns a current magnitude bound for the generators.

PowerModelsDistribution._calc_ground_shunt_admittance_matrix — Method

Given a set of terminals 'cnds' with associated shunt admittance 'Y', this method will calculate the reduced admittance matrix if terminal 'ground' is grounded.

PowerModelsDistribution._calc_load_current_magnitude_bounds — Method

Returns magnitude bounds for the current going through the load.

PowerModelsDistribution._calc_load_current_max — Method

Returns a magnitude bound for the current going through the load.

PowerModelsDistribution._calc_load_pq_bounds — Method

Calculates lower and upper bounds for the loads themselves (not the power withdrawn at the bus).

PowerModelsDistribution._calc_load_vbounds — Method

Returns the voltage magnitude bounds for the individual load elements in a multiphase load. These are inferred from vmin/vmax for wye loads and from calcbusvmll_bounds for delta loads.

PowerModelsDistribution._calc_max_cost_index — Method

_calc_max_cost_index(data::Dict{String,<:Any})

Computes maximum cost index of subnetworks

PowerModelsDistribution._calc_mc_transformer_Tvi — Method

calculates transformer Tvi

PowerModelsDistribution._calc_mc_voltage_product_bounds — Method

PowerModelsDistribution._calc_shunt — Method

Given a set of addmittances 'y' connected from the conductors 'fcnds' to the conductors 'tcnds', this method will return a list of conductors 'cnd' and a matrix 'Y', which will satisfy I[cnds] = Y*V[cnds].

PowerModelsDistribution._calc_transformer_current_max_frto — Method

Returns a current magnitude bound for the from and to side of a transformer. The total power rating also implies a current bound through the lower bound on the voltage magnitude of the connected buses.

PowerModelsDistribution._calc_transformer_power_ub_frto — Method

Returns a power magnitude bound for the from and to side of a transformer. The total current rating also implies a current bound through the upper bound on the voltage magnitude of the connected buses.

PowerModelsDistribution._cc_dfs — Method

DFS on a graph

PowerModelsDistribution._check_branch_loops — Method

checks that all branches connect two distinct buses

PowerModelsDistribution._check_bus — Method

bus data checks

PowerModelsDistribution._check_bus_and_terminals — Function

checks bus_name exists and has terminals

PowerModelsDistribution._check_configuration_infer_dim — Method

checks the connection configuration and infers the dimensions of the connection (number of connected terminals)

PowerModelsDistribution._check_connectivity — Method

checks connectivity of object

PowerModelsDistribution._check_connectivity — Method

checks that all buses are unique and other components link to valid buses

PowerModelsDistribution._check_eng_component_dtypes — Method

checks that an engineering model component has the correct data types

PowerModelsDistribution._check_equal — Method

checks if data structures are equivalent, and if not, will enumerate the differences

PowerModelsDistribution._check_generator — Method

generator data checks

PowerModelsDistribution._check_has_keys — Method

checks that a component has fields

PowerModelsDistribution._check_has_size — Method

check that fields has size data_size

PowerModelsDistribution._check_line — Method

line data checks

PowerModelsDistribution._check_linecode — Method

linecode data checks

PowerModelsDistribution._check_load — Method

load data checks

PowerModelsDistribution._check_load_needs_cone — Method

Returns a Bool, indicating whether the convex hull of the voltage-dependent relationship needs a cone inclusion constraint.

PowerModelsDistribution._check_same_size — Method

check that all data in fields have the same size

PowerModelsDistribution._check_shunt — Method

shunt data checks

PowerModelsDistribution._check_shunt_capacitor — Method

shunt capacitor data checks

PowerModelsDistribution._check_transformer — Method

Transformer, n-windings three-phase lossy data checks

PowerModelsDistribution._check_var_keys — Method

checks if a sufficient number of variables exist for the given keys collection

PowerModelsDistribution._check_voltage_source — Method

voltage source data checks

PowerModelsDistribution._clean_arrays! — Method

cleans up array properties back into strings for later conversion

PowerModelsDistribution._collect_nw_bus_lookups! — Method

PowerModelsDistribution._collect_nw_maps! — Method

PowerModelsDistribution._compute_bus_type — Method

computes the bus type based on existing bustype, the status of the generation object, and the controlmode

PowerModelsDistribution._convert_grounding — Method

lossy grounding to perfect grounding and shunts

PowerModelsDistribution._correct_branch_directions! — Method

checks that all parallel branches have the same orientation

PowerModelsDistribution._correct_bus_types! — Method

checks bus types are suitable for a power flow study, if not, fixes them. the primary checks are that all type 2 buses (i.e., PV) have a connected and active generator and there is a single type 3 bus (i.e., slack bus) with an active connected generator. Assumes that the network is a single connected component

PowerModelsDistribution._correct_cost_function! — Method

throws warnings if cost functions are malformed

PowerModelsDistribution._correct_cost_functions! — Method

throws warnings if cost functions are malformed

PowerModelsDistribution._correct_mc_thermal_limits! — Method

_correct_mc_thermal_limits!(data::Dict{String,<:Any})

checks that each branch has non-negative thermal ratings and removes zero thermal ratings

PowerModelsDistribution._correct_mc_voltage_angle_differences! — Function

_correct_mc_voltage_angle_differences!(data::Dict{String,<:Any}, default_pad::Real=deg2rad(10.0))

checks that voltage angle differences are within 90 deg., if not tightens to a default of 10deg (adjustable)

PowerModelsDistribution._create_capacitor — Function

Creates a Dict{String,Any} containing all of the expected properties for a Capacitor. If bus2 is not specified, the capacitor will be treated as a shunt. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_capcontrol — Function

Creates a Dict{String,Any} containing all expected properties for a CapControl object. See OpenDSS documentation for valid fields and ways to specify different properties.

PowerModelsDistribution._create_circuit — Function

alias createcircuit to createvsource

PowerModelsDistribution._create_generator — Function

Creates a Dict{String,Any} containing all of the expected properties for a Generator. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_growthshape — Function

Creates a Dict{String,Any} containing all expected properties for a GrowthShape element. See OpenDSS documentation for valid fields and ways to specify different properties.

PowerModelsDistribution._create_line — Function

Creates a Dict{String,Any} containing all of the properties for a Line. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_linecode — Function

Creates a Dict{String,Any} containing all of the properties of a Linecode. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_load — Function

Creates a Dict{String,Any} containing all of the expected properties for a Load. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_loadshape — Function

Creates a Dict{String,Any} containing all expected properties for a LoadShape element. See OpenDSS documentation for valid fields and ways to specify different properties.

PowerModelsDistribution._create_options — Method

PowerModelsDistribution._create_pvsystem — Function

Creates a Dict{String,Any} containing all of the expected properties for a PVSystem. See OpenDSS document https://github.com/tshort/OpenDSS/blob/master/Doc/OpenDSS%20PVSystem%20Model.doc for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_reactor — Function

Creates a Dict{String,Any} containing all of the expected properties for a Reactor. If bus2 is not specified Reactor is treated like a shunt. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_regcontrol — Function

Creates a Dict{String,Any} containing all expected properties for a RegControl object. See OpenDSS documentation for valid fields and ways to specify different properties.

PowerModelsDistribution._create_spectrum — Function

Creates a Dict{String,Any} containing all expected properties for a Spectrum object. See OpenDSS documentation for valid fields and ways to specify different properties.

PowerModelsDistribution._create_storage — Function

Creates a Dict{String,Any} containing all expected properties for a storage element. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_transformer — Function

Creates a Dict{String,Any} containing all of the expected properties for a Transformer. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_vsource — Function

Creates a Dict{String,Any} containing all of the expected properties for a Voltage Source. If bus2 is not specified, VSource will be treated like a generator. Mostly used as source which represents the circuit. See OpenDSS documentation for valid fields and ways to specify the different properties.

PowerModelsDistribution._create_xfmrcode — Function

Transformer codes contain all of the same properties as a transformer except bus, buses, bank, xfmrcode

PowerModelsDistribution._create_xycurve — Function

Creates a Dict{String,Any} containing all expected properties for a XYCurve object. See OpenDSS documentation for valid fields and ways to specify different properties.

PowerModelsDistribution._discover_buses — Method

Discovers all of the buses (not separately defined in OpenDSS), from 'lines'

PowerModelsDistribution._discover_phases_neutral! — Method

discovers all phases and neutrals in the network

PowerModelsDistribution._discover_terminals! — Method

discovers all terminals in the network

PowerModelsDistribution._dss2eng_bus! — Function

Adds nodes as buses to data_eng from data_dss

PowerModelsDistribution._dss2eng_buscoords! — Method

Parses buscoords lon,lat (if present) into their respective buses

PowerModelsDistribution._dss2eng_capacitor! — Method

Adds capacitors to data_eng from data_dss

PowerModelsDistribution._dss2eng_capcontrol! — Method

Adds capcontrol to data_eng from data_dss

PowerModelsDistribution._dss2eng_generator! — Function

Adds generators to data_eng from data_dss

PowerModelsDistribution._dss2eng_line! — Method

Adds lines to data_eng from data_dss

PowerModelsDistribution._dss2eng_linecode! — Method

Adds lines to data_eng from data_dss

PowerModelsDistribution._dss2eng_load! — Function

Adds loads to data_eng from data_dss

Constant can still be scaled by other settings, fixed cannot Note that in the current feature set, fixed therefore equals constant

1: Constant P and Q, default 2: Constant Z 3: Constant P and quadratic Q 4: Exponential 5: Constant I 6: Constant P and fixed Q

7: Constant P and quadratic Q (i.e., fixed reactance)

8: ZIP

PowerModelsDistribution._dss2eng_loadshape! — Function

Adds loadshapes to data_eng from data_dss

PowerModelsDistribution._dss2eng_pvsystem! — Function

Adds pvsystems to data_eng from data_dss

PowerModelsDistribution._dss2eng_reactor! — Method

Adds shunt reactors to data_eng from data_dss

PowerModelsDistribution._dss2eng_regcontrol! — Method

Adds regcontrol to data_eng from data_dss

PowerModelsDistribution._dss2eng_storage! — Function

Adds storage to data_eng from data_dss

PowerModelsDistribution._dss2eng_transformer! — Method

Adds transformers to data_eng from data_dss

PowerModelsDistribution._dss2eng_vsource! — Function

Adds vsources to data_eng from data_dss

PowerModelsDistribution._dss2eng_xfmrcode! — Method

Adds transformers to data_eng from data_dss

PowerModelsDistribution._find_neutrals — Method

Discovers all neutrals in the network

PowerModelsDistribution._find_ref_buses — Method

returns all reference buses (bus_type==3)

PowerModelsDistribution._fix_arrays! — Method

helper function to fix matrices (from vector of vectors) and vector dtypes

PowerModelsDistribution._fix_dtypes! — Method

recursive function to fix data types from data imported from json

PowerModelsDistribution._fix_enums! — Method

helper function to convert stringified enums

PowerModelsDistribution._fix_nulls! — Method

helper function to fix null values from json (usually Inf or NaN)

PowerModelsDistribution._generate_vm_pairs — Method

Generates pairwise bounds for oneport components.

PowerModelsDistribution._get_complete_conductor_set — Method

finds maximal set of ungrounded phases

PowerModelsDistribution._get_conductor_indicator — Method

returns the conductor indicator for a ENGINEERING component

PowerModelsDistribution._get_conductors_ordered — Method

Returns an ordered list of defined conductors. If ground=false, will omit any 0

PowerModelsDistribution._get_delta_transformation_matrix — Method

creates a delta transformation matrix

PowerModelsDistribution._get_ground! — Method

gets the grounding information for a bus

PowerModelsDistribution._get_ground_math! — Method

get a grounded terminal from a bus; if not present, create one

PowerModelsDistribution._get_idxs — Method

Given a vector and a list of elements to find, this method will return a list of the positions of the elements in that vector.

PowerModelsDistribution._get_ilocs — Method

get locations of terminal in connections list

PowerModelsDistribution._get_implied_nphases — Method

returns number of phases implied by a two-bus (edge) object

PowerModelsDistribution._get_implied_nphases — Method

returns number of phases implied by a single-bus (node) object

PowerModelsDistribution._get_implied_nphases — Method

returns number of phases implied by a transformer object

PowerModelsDistribution._get_line_impedance_parameters — Method

Obtain impedance parameters, directly or from linecode.

PowerModelsDistribution._get_math_obj — Method

returns component from the mathematical data model

PowerModelsDistribution._get_new_ground — Method

getground helper function

PowerModelsDistribution._get_tight_absolute_voltage_magnitude_bounds — Method

Returns the tightest set of absolute voltage magnitude bounds, removing looser bounds which are implied by the tighter ones.

PowerModelsDistribution._get_tight_pairwise_voltage_magnitude_bounds — Method

Returns the tightest set of pairwise voltage magnitude bounds, removing looser bounds which are implied by the tighter ones.

PowerModelsDistribution._guess_dtype — Method

guesses the data type of a value using regex, returning Float64, Int, ComplexF64, or String (if number type cannot be determined)

PowerModelsDistribution._has_nl_expression — Method

helper to determine if expession has any Nonlinear terms

PowerModelsDistribution._impedance_conversion — Method

converts impendance in Ohm/m by multiplying by length

PowerModelsDistribution._import_all! — Method

creates a dss dict inside object that imports all items in prop_order from dss_obj

PowerModelsDistribution._infer_int_dim — Method

infer the internal dimension of a winding, load or generator based on the connections and the configuration

PowerModelsDistribution._infer_int_dim_transformer — Method

infer the internal dimension for a transformer (only in the MATHEMATICAL data model format)

PowerModelsDistribution._infer_int_dim_unit — Method

infer the internal dimension for a unit, i.e. any one-port component with connections and configuration properties

PowerModelsDistribution._infer_neutral_terminals — Method

Return a list of all implicit neutrals as a list of bus-terminal pairs. This is done by starting from a list of all terminals which are either a.connected to the neutral of wye-connected components; b. or are grounded. This initial list is then expanded to all terminals which are galvanically connected to terminals in the initial list.

PowerModelsDistribution._init_base_components! — Method

initializes the base components that are expected by powermodelsdistribution in the mathematical model

PowerModelsDistribution._init_lookup! — Method

Initializes the lookup table

PowerModelsDistribution._init_math_obj — Method

initializes the base math object of any type, and copies any one-to-one mappings

PowerModelsDistribution._init_unmap_eng_obj! — Method

initialization actions for unmapping

PowerModelsDistribution._is_after — Method

checks to see if property1 is after property2 in the prop_order

PowerModelsDistribution._is_after_linecode — Method

checks to see if a property is after linecode

PowerModelsDistribution._is_after_xfmrcode — Method

checks to see if a property is after xfmrcode

PowerModelsDistribution._is_loadshape_split — Method

checks if loadshape has both pmult and qmult

PowerModelsDistribution._is_multiport_component — Method

Indicates whether the passed component has a multiport structure (e.g. transformers).

PowerModelsDistribution._is_oneport_component — Method

Indicates whether the passed component has a oneport structure (e.g. loads and generators).

PowerModelsDistribution._is_twoport_component — Method

Indicates whether the passed component has a twoport structure (e.g. lines and switches).

PowerModelsDistribution._isa_array — Method

checks if data is an opendss-style array string

PowerModelsDistribution._isa_conn — Method

checks is a string is a connection by checking the values

PowerModelsDistribution._isa_matrix — Method

checks if data is an opendss-style matrix string

PowerModelsDistribution._isa_rpn — Method

detects if expr is Reverse Polish Notation expression

PowerModelsDistribution._kron_reduce_branch! — Method

performs kron reduction on branch

PowerModelsDistribution._kron_reduce_branch — Method

performs kron reduction on branch - helper function

PowerModelsDistribution._kron_reduce_buses! — Method

transformations might have introduced buses with four-terminals; crop here

PowerModelsDistribution._kron_reduce_implicit_neutrals! — Method

_kron_reduce_implicit_neutrals!(data_eng::Dict{String,Any})::Dict{String,Any}

Kron-reduce all (implied) neutral conductors of lines, switches and shunts, and remove any terminals which become unconnected. A line or switch conductor is considered as a neutral conductor if it is connected between two neutral terminals. A terminal is a neutral terminals if it is galvanically connected (i.e. through a line or switch) to a grounded terminal, or the neutral conductor of a wye-connected component.

PowerModelsDistribution._kron_reduce_linecode! — Method

Kron-reduce specified neutral conductors of a linecode.

PowerModelsDistribution._kron_reduce_series_impedance — Method

Return the Kron-reduction of the specified neutral conductors of a series impedance matrix.

PowerModelsDistribution._kron_reduce_shunt_addmittance — Method

Return the Kron-reduction of the specified neutral conductors of a shunt addmittance matrix.

PowerModelsDistribution._line_reverse! — Method

Reverse the direction of a line.

PowerModelsDistribution._load_expmodel_params — Method

_load_expmodel_params(load::Dict{String,<:Any}, bus::Dict{String,<:Any})

Returns the exponential load model parameters for a load. For an exponential load it simply returns certain data model properties, whilst for constantpower, constantcurrent and constant_impedance it returns the equivalent exponential model parameters.

PowerModelsDistribution._loop_line_to_shunt — Method

Create an equivalent shunt for a line which connects to a single bus.

PowerModelsDistribution._make_filtered_logger — Method

_make_filtered_logger(level::Logging.LogLevel)

Helper function to create the filtered logger for PMD

PowerModelsDistribution._make_full_matrix_variable — Method

makes a full matrix variable from a diagonal, and lower and upper triangular vectors

PowerModelsDistribution._make_hermitian_matrix_variable — Method

makes a hermitian matrix variable from diagonal, and lower and upper triangular vectors

PowerModelsDistribution._make_lossless! — Method

_make_lossless!(data_eng::Dict{String,<:Any})

Remove parameters from objects with loss models to make them lossless. This includes linecodes, lines, switches, xfmrcodes, transformers, voltage sources, generators, solar, and storage, which all have (or will have in the future), loss model parameters that can be omitted.

PowerModelsDistribution._make_math_per_unit! — Method

converts the MATHEMATICAL model to per unit from SI

PowerModelsDistribution._make_matrix_variable_element — Method

_make_matrix_variable_element(
    model::JuMP.Model,
    indices::Array{T,1},
    n::Int,
    m::Int;
    upper_bound::Union{Missing, Dict{T,<:Matrix{<:Real}}}=missing,
    lower_bound::Union{Missing, Dict{T,<:Matrix{<:Real}}}=missing,
    varname::String=""
) where T

Sometimes we want to bound only a subset of the elements of a matrix variable. For example, an unbounded Hermitian variable usually still has a lower bound of zero on the real diagonal elements. When there is a mix of bounded and unbounded elements, the unboundedness is encoded as 'Inf' and '-Inf' in the bound parameters. This cannot be passed directlty to JuMP, because it would lead to an error in Mosek for example. Instead, this method checks whether all bounds for an element (n,m) are Inf, and if so, does not pass a bound to JuMP.

PowerModelsDistribution._make_multiconductor! — Method

Hacky helper function to transform single-conductor network data, from, e.g., matpower/psse, into multi-conductor data

PowerModelsDistribution._make_multinetwork_eng — Method

Expands an ENGINEERING data structure into a multinetwork, see make_multinetwork

PowerModelsDistribution._make_multinetwork_math — Method

Expands an MATHEMATICAL data structure into a multinetwork, see make_multinetwork

PowerModelsDistribution._map_conductor_ids! — Method

helper function to map non integer conductor ids into integers

PowerModelsDistribution._map_eng2math — Method

base function for converting engineering model to mathematical model

PowerModelsDistribution._map_eng2math_bus! — Method

converts engineering bus components into mathematical bus components

PowerModelsDistribution._map_eng2math_generator! — Method

converts engineering generators into mathematical generators

PowerModelsDistribution._map_eng2math_line! — Method

converts engineering lines into mathematical branches

PowerModelsDistribution._map_eng2math_load! — Method

converts engineering load components into mathematical load components

PowerModelsDistribution._map_eng2math_nw! — Method

PowerModelsDistribution._map_eng2math_shunt! — Method

converts engineering generic shunt components into mathematical shunt components

PowerModelsDistribution._map_eng2math_solar! — Method

converts engineering solar components into mathematical generators

PowerModelsDistribution._map_eng2math_storage! — Method

converts engineering storage into mathematical storage

PowerModelsDistribution._map_eng2math_switch! — Method

converts engineering switches into mathematical switches and (if neeed) impedance branches to represent loss model

PowerModelsDistribution._map_eng2math_transformer! — Method

converts engineering n-winding transformers into mathematical ideal 2-winding lossless transformer branches and impedance branches to represent the loss model

PowerModelsDistribution._map_eng2math_voltage_source! — Method

converts engineering voltage sources into mathematical generators and (if needed) impedance branches to represent the loss model

PowerModelsDistribution._map_math2eng_bus! — Method

PowerModelsDistribution._map_math2eng_generator! — Method

PowerModelsDistribution._map_math2eng_line! — Method

PowerModelsDistribution._map_math2eng_load! — Method

PowerModelsDistribution._map_math2eng_root! — Method

PowerModelsDistribution._map_math2eng_shunt! — Method

PowerModelsDistribution._map_math2eng_solar! — Method

PowerModelsDistribution._map_math2eng_storage! — Method

PowerModelsDistribution._map_math2eng_switch! — Method

PowerModelsDistribution._map_math2eng_transformer! — Method

PowerModelsDistribution._map_math2eng_voltage_source! — Method

PowerModelsDistribution._mat2ltrivec! — Method

matrix to lower triangular vector

PowerModelsDistribution._mat2utrivec! — Method

matrix to upper triangular vector

PowerModelsDistribution._mat_mult_rm_nan — Method

PowerModelsDistribution._merge_bus_flows — Method

Merges flow variables that enter the same terminals, i.e. multiple neutrals of an underground cable connected to same neutral terminal

PowerModelsDistribution._merge_dss! — Method

Merges two (partially) parsed OpenDSS files to the same dictionary dss_prime. Used in cases where files are referenced via the "compile" or "redirect" OpenDSS commands inside the originating file.

PowerModelsDistribution._merge_terminals! — Method

Merge a terminal into another for a specified bus, i.e. as if they are short-ciruited.

PowerModelsDistribution._missing2false — Method

Helper to convert

PowerModelsDistribution._nan2zero — Method

Replaces NaN values with zeros

PowerModelsDistribution._new_terminal — Method

generate a new, unique terminal

PowerModelsDistribution._objective_mc_min_fuel_cost_polynomial_linquad — Method

gen connections adaptation of min fuel cost polynomial linquad objective

PowerModelsDistribution._objective_mc_min_fuel_cost_polynomial_linquad_switch — Method

Multiconductor adaptation of min fuel cost polynomial linquad objective

PowerModelsDistribution._objective_mc_min_fuel_cost_polynomial_linquad_switch — Method

gen connections adaptation of min fuel cost polynomial linquad objective

PowerModelsDistribution._objective_mc_min_fuel_cost_polynomial_nl — Method

PowerModelsDistribution._objective_mc_min_fuel_cost_polynomial_nl_switch — Method

PowerModelsDistribution._pad_connections! — Method

adds conductors to connections during padding process, transformer winding variant

PowerModelsDistribution._pad_connections! — Method

adds conductors to connections during padding process

PowerModelsDistribution._pad_properties! — Method

pads properties to have the total number of conductors for the whole system (transformer winding variant)

PowerModelsDistribution._pad_properties! — Method

pads properties to have the total number of conductors for the whole system

PowerModelsDistribution._pad_properties_delta! — Method

pads properties to have the total number of conductors for the whole system - delta connection variant

PowerModelsDistribution._pad_properties_delta! — Method

pads properties to have the total number of conductors for the whole system - delta connection variant

PowerModelsDistribution._parse_array — Method

Parses a OpenDSS style array string data into a one dimensional array of type dtype. Array strings are capped by either brackets, single quotes, or double quotes, and elements are separated by spaces.

PowerModelsDistribution._parse_binary_file — Method

parses sng and dbl precision loadshape binary files

PowerModelsDistribution._parse_bus_id — Method

Parses busnames as defined in OpenDSS, e.g. 'primary.1.2.3.0'

PowerModelsDistribution._parse_buscoords_file — Method

Parses a Bus Coordinate file, in either "dat" or "csv" formats, where in "dat", columns are separated by spaces, and in "csv" by commas. File expected to contain "bus,x,y" on each line.

PowerModelsDistribution._parse_component — Function

Parses a component with properties into a object. If object is not defined, an empty dictionary will be used. Assumes that unnamed properties are given in order, but named properties can be given anywhere.

PowerModelsDistribution._parse_conn — Method

parses connection 'conn' specification reducing to wye or delta

PowerModelsDistribution._parse_csv_file — Method

parses single column load profile files

PowerModelsDistribution._parse_data_file — Method

parses csv or binary loadshape files

PowerModelsDistribution._parse_dss_capcontrol_type! — Method

_parse_dss_capcontrol_type!(type::SubString{String}, id::Any)

Converts dss capcontrol type to supported PowerModelsDistribution CapControlType enum.

PowerModelsDistribution._parse_dss_load_model! — Method

converts dss load model to supported PowerModelsDistribution LoadModel enum

PowerModelsDistribution._parse_dss_with_dtypes! — Function

Parses the data in keys defined by to_parse in data_dss using types given by the default properties from the get_prop_default function.

PowerModelsDistribution._parse_dss_xycurve — Method

PowerModelsDistribution._parse_element_with_dtype — Method

parses the raw dss values into their expected data types

PowerModelsDistribution._parse_line — Method

Parses an already separated line given by elements (an array) of an OpenDSS file into current_obj. If not defined, current_obj is an empty dictionary.

PowerModelsDistribution._parse_loadshape! — Method

parses loadshape component

PowerModelsDistribution._parse_matrix — Method

Parses a OpenDSS style triangular matrix string data into a two dimensional array of type dtype. Matrix strings are capped by either parenthesis or brackets, rows are separated by "|", and columns are separated by spaces.

PowerModelsDistribution._parse_mult_parameter — Method

parses pmult and qmult entries on loadshapes

PowerModelsDistribution._parse_obj_dtypes! — Method

parses data type of properties of objects

PowerModelsDistribution._parse_properties — Method

Parses a string of properties of a component type, character by character into an array with each element containing (if present) the property name, "=", and the property value.

PowerModelsDistribution._parse_rpn — Function

parses Reverse Polish Notation expr

PowerModelsDistribution._parse_setbusxy! — Method

parses the setbusxy command

PowerModelsDistribution._parse_spectrum! — Method

parse spectrum component

PowerModelsDistribution._parse_xycurve! — Method

parse xycurve component

PowerModelsDistribution._pmd_metafmt — Method

_pmd_metafmt(level::Logging.LogLevel, _module, group, id, file, line)

MetaFormatter for ConsoleLogger for PMD to adjust log message format

PowerModelsDistribution._propagate_network_topology! — Method

_propagate_network_topology!(data::Dict{String,Any})

helper function to propagate bus status to any connected components

PowerModelsDistribution._rebase_pu_branch! — Method

per-unit conversion for branches

PowerModelsDistribution._rebase_pu_bus! — Method

per-unit conversion for buses

PowerModelsDistribution._rebase_pu_generator! — Method

per-unit conversion for generators

PowerModelsDistribution._rebase_pu_load! — Method

per-unit conversion for loads

PowerModelsDistribution._rebase_pu_shunt! — Method

per-unit conversion for shunts

PowerModelsDistribution._rebase_pu_storage! — Method

per-unit conversion for storage

PowerModelsDistribution._rebase_pu_switch! — Method

per-unit conversion for switches

PowerModelsDistribution._rebase_pu_transformer_2w_ideal! — Method

per-unit conversion for ideal 2-winding transformers

PowerModelsDistribution._ref_add_connected_components! — Method

adds connected components for opb problem type

PowerModelsDistribution._register_awaiting_ground! — Method

PowerModelsDistribution._remove_all_bounds! — Method

_remove_all_bounds!(data_eng; exclude::Vector{<:String}=String["energy_ub"], exclude_asset_type::Vector{String}=String[])

Removes all fields ending in 'ub' or 'lb' that aren't required by the math model. Properties can be excluded from this removal with exclude::Vector{String}

Whole asset types (e.g., "line") can be excluded using the keyword argument exclude_asset_type::Vector{String}

By default, "energy_ub" is excluded from this removal, since it is a required properly on storage.

PowerModelsDistribution._remove_line_limits! — Method

_remove_line_limits!(data_eng::Dict{String,<:Any})

Removes fields cm_ub and sm_ub from lines, switches, and linecodes

PowerModelsDistribution._remove_pwl_cost_duplicates! — Method

checks that each point in the a pwl function is unqiue, simplifies the function if duplicates appear

PowerModelsDistribution._remove_transformer_limits! — Method

_remove_transformer_limits!(data_eng::Dict{String,<:Any})

Removes field sm_ub from transformers, xfmrcodes

PowerModelsDistribution._remove_unconnected_terminals! — Method

_remove_unconnected_terminals!(data_eng::Dict{String,Any})::Dict{String,Any}

Remove all terminals which are unconnected (not considering a grounding as a connection).

PowerModelsDistribution._replace_nan — Method

Replaces NaN values with zeros

PowerModelsDistribution._rescale_cost_model! — Method

rescales the cost model terms

PowerModelsDistribution._roll — Method

rolls a 1d array left or right by idx

PowerModelsDistribution._sc2br_impedance — Method

Converts a set of short-circuit tests to an equivalent reactance network. Reference: R. C. Dugan, “A perspective on transformer modeling for distribution system analysis,” in 2003 IEEE Power Engineering Society General Meeting (IEEE Cat. No.03CH37491), 2003, vol. 1, pp. 114-119 Vol. 1.

PowerModelsDistribution._scale — Method

function for applying a scale to a paramter

PowerModelsDistribution._scale_props! — Method

helper function to apply a scale factor to given properties

PowerModelsDistribution._simplify_cost_terms! — Method

PowerModelsDistribution._simplify_pwl_cost! — Method

checks the slope of each segment in a pwl function, simplifies the function if the slope changes is below a tolerance

PowerModelsDistribution._slice_branches! — Method

slices branches based on connected terminals

PowerModelsDistribution._sol_data_model_acr! — Method

_sol_data_model_acr!(solution::Dict{String,<:Any})

solutionprocessor, see [`solvemcmodel`](@ref solvemc_model), to convert ACR variables back into polar representation (default data model voltage form)

PowerModelsDistribution._sol_data_model_w! — Method

converts w models voltages to standard voltage magnitude (sqrt)

PowerModelsDistribution._solve_mc_model — Method

_solve_mc_model(
    data::Dict{String,<:Any},
    model_type::Type,
    optimizer,
    build_method::Function;
    multinetwork::Bool=false,
    ref_extensions::Vector{<:Function}=Function[],
    solution_processors::Vector{<:Function}=Function[],
    relax_integrality::Bool=false,
    kwargs...
)::Dict{String,Any}

Internal solver interface that uses instantiate_mc_model directly and runs optimize_model!, returning a result

See solve_mc_model

PowerModelsDistribution._solve_mc_osw — Method

Solve optimal switching problem

PowerModelsDistribution._solve_mc_osw_mi — Method

Solve mixed-integer optimal switching problem

PowerModelsDistribution._solve_mn_mc_opb — Method

_solve_mn_mc_opb(data::Union{Dict{String,<:Any},String}, model_type::Type, solver; kwargs...)

solve test mn mc problem

PowerModelsDistribution._standardize_cost_terms! — Method

ensures all polynomial costs functions have at exactly comp_order terms

PowerModelsDistribution._strip_comments — Method

Strips comments, defined by '!' from the ends of lines

PowerModelsDistribution._strip_lines — Method

strips lines that are either commented (block or single) or empty

PowerModelsDistribution._sum_rm_nan — Method

BOUND manipulation methods (0*Inf->0 is often desired)

PowerModelsDistribution._to_kwargs — Method

converts Dict{String,Any} to Dict{Symbol,Any} for passing as kwargs

PowerModelsDistribution._transform_loops! — Method

_transform_loops!(
    data_eng::Dict{String,Any};
    zero_series_impedance_threshold::Real=1E-8,
    shunt_id_prefix::AbstractString="line_loop"
)::Dict{String,Any}

Transform line loops (connected to a single bus), which are not allowed in the mathematical model. Lossy line loops are converted to equivalent shunts, and lossless ones (i.e. short-circuits) are represented by merging the short-circuited terminals. The argument 'zeroseriesimpedance_threshold' controls the threshold below which the series impedance is considered to be a short-ciruit. This is useful because OpenDSS modelers have to insert tiny impedances to represent short-circuit reactors. The addmittance to ground should be zero to trigger the short-circuit handling.

PowerModelsDistribution._update_bus_terminal_projections! — Method

helper function to update the terminals on projected buses

PowerModelsDistribution._vec2ltri! — Method

vector to lower triangular

PowerModelsDistribution._vec2utri! — Method

https://stackoverflow.com/questions/39039553/lower-triangular-matrix-in-julia

PowerModelsDistribution._wrap_to_180 — Method

wraps angles in degrees to 180

PowerModelsDistribution._wrap_to_pi — Method

wraps angles in radians to pi