Objectives

PowerModelsDistribution.objective_mc_max_load_setpoint — Method

objective_mc_max_load_setpoint(pm::AbstractUnbalancedPowerModel)

maximum loadability objective (continuous load shed) with storage

PowerModelsDistribution.objective_mc_min_fuel_cost — Method

objective_mc_min_fuel_cost(pm::AbstractUnbalancedPowerModel)

Standard fuel cost minimization objective

PowerModelsDistribution.objective_mc_min_fuel_cost_polynomial — Method

objective_mc_min_fuel_cost_polynomial(pm::AbstractUnbalancedPowerModel)

Fuel cost minimization objective for polynomial terms

PowerModelsDistribution.objective_mc_min_fuel_cost_polynomial_switch — Method

objective_mc_min_fuel_cost_polynomial_switch(pm::AbstractUnbalancedPowerModel)

Fuel cost minimization objective for polynomial terms including switches

PowerModelsDistribution.objective_mc_min_fuel_cost_pwl — Method

objective_mc_min_fuel_cost_pwl(pm::AbstractUnbalancedPowerModel)

Fuel cost minimization objective with piecewise linear terms

PowerModelsDistribution.objective_mc_min_fuel_cost_pwl_switch — Method

objective_mc_min_fuel_cost_pwl_switch(pm::AbstractUnbalancedPowerModel)

Fuel cost minimization objective with piecewise linear terms including switches

PowerModelsDistribution.objective_mc_min_fuel_cost_switch — Method

objective_mc_min_fuel_cost_switch(pm::AbstractUnbalancedPowerModel)

Standard fuel cost minimization objective including switches

PowerModelsDistribution.objective_mc_min_load_setpoint_delta — Method

objective_mc_min_load_setpoint_delta(pm::AbstractUnbalancedPowerModel)

minimum load delta objective with storage

PowerModelsDistribution.objective_mc_min_load_setpoint_delta_simple — Method

objective_mc_min_load_setpoint_delta_simple(pm::AbstractUnbalancedPowerModel)

simplified minimum load delta objective (continuous load shed)

PowerModelsDistribution.objective_mc_min_load_setpoint_delta_simple_switch — Method

objective_mc_min_load_setpoint_delta_simple_switch(pm::AbstractUnbalancedPowerModel)

simplified minimum load delta objective (continuous load shed) including a switch state term

PowerModelsDistribution.objective_mc_min_slack_bus_power — Method

objective_mc_min_slack_bus_power(pm::AbstractUnbalancedPowerModel)

a quadratic penalty for bus power slack variables

PowerModelsDistribution.objective_mc_variable_pg_cost — Method

objective_mc_variable_pg_cost(pm::AbstractUnbalancedPowerModel)

adds pg_cost variables and constraints

PowerModelsDistribution.objective_variable_pg_cost — Method

objective_variable_pg_cost(pm::AbstractUnbalancedIVRModel)

adds pg_cost variables and constraints for the IVR formulation

Helpers

PowerModelsDistribution.calc_max_cost_index — Function

calc_max_cost_index(data::Dict{String,<:Any})

Computes maximum cost index

PowerModelsDistribution.simplify_cost_terms! — Function

simplify_cost_terms!(data::Dict{String,<:Any})

trims zeros from higher order cost terms

PowerModelsDistribution.calc_pwl_points — Function

calc_pwl_points(ncost::Int, cost::Vector{<:Real}, pmin::Real, pmax::Real; tolerance=1e-2)

cleans up raw pwl cost points in preparation for building a mathamatical model. The key mathematical properties,

the first and last points are strickly outside of the pmin-to-pmax range

pmin and pmax occur in the first and last line segments.

PowerModelsDistribution.calc_cost_pwl_lines — Function

calc_cost_pwl_lines(comp_dict::Dict)

compute lines in m and b from from pwl cost models data is a list of components. Can be run on data or ref data structures

PowerModelsDistribution.standardize_cost_terms! — Function

standardize_cost_terms!(data::Dict{String,<:Any}; order=-1)

ensures all polynomial costs functions have the same number of terms