Modeling Components

ThreePhasePowerModels.jl Library

ThreePhasePowerModels.LPLinUBFForm — Type.

LinDist3Flow per Sankur et al 2016, using vector variables for power, voltage and current in scalar form

source

ThreePhasePowerModels.LPLinUBFPowerModel — Type.

source

ThreePhasePowerModels.LPLinUBFPowerModel — Method.

default Lin3Distflow constructor for scalar form

source

ThreePhasePowerModels.LPdiagUBFForm — Type.

LinDist3Flow per Sankur et al 2016, using vector variables for power, voltage and current

source

ThreePhasePowerModels.LPdiagUBFPowerModel — Type.

source

ThreePhasePowerModels.LPdiagUBFPowerModel — Method.

default LP unbalanced DistFlow constructor

source

ThreePhasePowerModels.LPfullUBFForm — Type.

Simplified BFM per Gan and Low 2014, PSCC, using matrix variables for power, voltage and current

source

ThreePhasePowerModels.LPfullUBFPowerModel — Type.

source

ThreePhasePowerModels.LPfullUBFPowerModel — Method.

default LP unbalanced DistFlow constructor

source

ThreePhasePowerModels.SDPUBFForm — Type.

SDP BFM per Gan and Low 2014, PSCC

source

ThreePhasePowerModels.SDPUBFPowerModel — Type.

source

ThreePhasePowerModels.SDPUBFPowerModel — Method.

default SDP unbalanced DistFlow constructor

source

ThreePhasePowerModels.SOCConicUBFForm — Type.

SOC relaxation of SDPUBFForm per Kim, Kojima, & Yamashita 2003, cast as a SOC

source

ThreePhasePowerModels.SOCConicUBFPowerModel — Type.

source

ThreePhasePowerModels.SOCConicUBFPowerModel — Method.

default SOC unbalanced DistFlow constructor

source

ThreePhasePowerModels.SOCNLPUBFForm — Type.

SOC relaxation of SDPUBFForm per Kim, Kojima, & Yamashita 2003, cast as an QCP

source

ThreePhasePowerModels.SOCNLPUBFPowerModel — Type.

source

ThreePhasePowerModels.SOCNLPUBFPowerModel — Method.

default SOC unbalanced DistFlow constructor

source

ThreePhasePowerModels.run_ac_tp_opf — Method.

source

ThreePhasePowerModels.run_ac_tp_pf — Method.

source

ThreePhasePowerModels.run_tp_opf — Method.

source

ThreePhasePowerModels.run_tp_opf — Method.

source

ThreePhasePowerModels.run_tp_opf_bf — Method.

source

ThreePhasePowerModels.run_tp_opf_bf — Method.

source

ThreePhasePowerModels.run_tp_opf_pbs — Method.

source

ThreePhasePowerModels.run_tp_opf_pbs — Method.

source

ThreePhasePowerModels.run_tp_ots — Method.

source

ThreePhasePowerModels.run_tp_ots — Method.

source

ThreePhasePowerModels.run_tp_pf — Method.

source

ThreePhasePowerModels.run_tp_pf — Method.

source

ThreePhasePowerModels.run_tp_pf_bf — Method.

source

ThreePhasePowerModels.run_tp_pf_bf — Method.

source

ThreePhasePowerModels.run_tp_pf_pbs — Method.

source

ThreePhasePowerModels.run_tp_pf_pbs — Method.

source

ThreePhasePowerModels.AbstractConicUBFForm — Type.

source

ThreePhasePowerModels.AbstractLPUBFForm — Type.

Abstract form for linear unbalanced power flow models

source

ThreePhasePowerModels.AbstractNLPUBFForm — Type.

source

ThreePhasePowerModels.DSSSourceId — Type.

Structure representing OpenDSS dss_source_id giving the type of the component dss_type, its name dss_name, and the active phases active_phases

source

PowerModels.constraint_branch_current — Method.

source

ThreePhasePowerModels.add_branch_current_setpoint — Method.

source

ThreePhasePowerModels.add_branch_flow_setpoint — Method.

source

ThreePhasePowerModels.add_bus_slack_setpoint — Method.

source

ThreePhasePowerModels.add_bus_voltage_setpoint — Method.

source

ThreePhasePowerModels.add_component! — Method.

add_component!(dss_data, ctype_name, compDict)

Adds a component of type ctype_name with properties given by compDict to the existing dss_data structure. If a component of the same type has already been added to dss_data, the new component is appeneded to the existing array of components of that type, otherwise a new array is created.

source

ThreePhasePowerModels.add_original_variables — Method.

source

ThreePhasePowerModels.add_property — Method.

add_property(compDict, key, value)

Adds a property to an existing component properties dictionary compDict given the key and value of the property. If a property of the same name already exists inside compDict, the original value is converted to an array, and the new value is appended to the end.

source

ThreePhasePowerModels.adjust_sourcegen_bounds! — Method.

adjust_sourcegen_bounds!(tppm_data)

Changes the bounds for the sourcebus generator by checking the emergamps of all of the branches attached to the sourcebus and taking the sum of non-infinite values. Defaults to Inf if all emergamps connected to sourcebus are also Inf.

source

ThreePhasePowerModels.assign_property! — Method.

assign_property!(dss_data, cType, cName, propName, propValue)

Assigns a property with name propName and value propValue to the component of type cType named cName in dss_data.

source

ThreePhasePowerModels.calc_tp_voltage_product_bounds — Method.

source

ThreePhasePowerModels.check_duplicate_components! — Method.

check_duplicate_components!(dss_data)

Finds duplicate components in dss_data and merges up, meaning that older data (lower indices) is always overwritten by newer data (higher indices).

source

ThreePhasePowerModels.check_network_data — Method.

source

ThreePhasePowerModels.constraint_flow_losses — Method.

Defines branch flow model power flow equations

source

ThreePhasePowerModels.constraint_kcl_shunt_slack — Method.

source

ThreePhasePowerModels.constraint_kcl_shunt_slack — Method.

source

ThreePhasePowerModels.constraint_kcl_shunt_slack — Method.

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from — Method.

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from — Method.

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p[f_idx] ==  (g+g_fr)/tm*v[f_bus]^2 + (-g*tr+b*ti)/tm*(v[f_bus]*v[t_bus]*cos(t[f_bus]-t[t_bus])) + (-b*tr-g*ti)/tm*(v[f_bus]*v[t_bus]*sin(t[f_bus]-t[t_bus]))
q[f_idx] == -(b+b_fr)/tm*v[f_bus]^2 - (-b*tr-g*ti)/tm*(v[f_bus]*v[t_bus]*cos(t[f_bus]-t[t_bus])) + (-g*tr+b*ti)/tm*(v[f_bus]*v[t_bus]*sin(t[f_bus]-t[t_bus]))

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from — Method.

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p[f_idx] == -b*(t[f_bus] - t[t_bus])

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from — Method.

nothing to do, no voltage angle variables

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from — Method.

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from_on_off — Method.

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from_on_off — Method.

p[f_idx] == z*(g/tm*v[f_bus]^2 + (-g*tr+b*ti)/tm*(v[f_bus]*v[t_bus]*cos(t[f_bus]-t[t_bus])) + (-b*tr-g*ti)/tm*(v[f_bus]*v[t_bus]*sin(t[f_bus]-t[t_bus])))
q[f_idx] == z*(-(b+c/2)/tm*v[f_bus]^2 - (-b*tr-g*ti)/tm*(v[f_bus]*v[t_bus]*cos(t[f_bus]-t[t_bus])) + (-g*tr+b*ti)/tm*(v[f_bus]*v[t_bus]*sin(t[f_bus]-t[t_bus])))

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from_on_off — Method.

-b*(t[f_bus] - t[t_bus] + vad_min*(1-branch_z[i])) <= p[f_idx] <= -b*(t[f_bus] - t[t_bus] + vad_max*(1-branch_z[i]))

source

ThreePhasePowerModels.constraint_ohms_tp_yt_from_on_off — Method.

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p[f_idx] ==        g/tm*w_fr[i] + (-g*tr+b*ti)/tm*(wr[i]) + (-b*tr-g*ti)/tm*(wi[i])
q[f_idx] == -(b+c/2)/tm*w_fr[i] - (-b*tr-g*ti)/tm*(wr[i]) + (-g*tr+b*ti)/tm*(wi[i])

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to — Method.

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to — Method.

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p[t_idx] ==  (g+g_to)*v[t_bus]^2 + (-g*tr-b*ti)/tm*(v[t_bus]*v[f_bus]*cos(t[t_bus]-t[f_bus])) + (-b*tr+g*ti)/tm*(v[t_bus]*v[f_bus]*sin(t[t_bus]-t[f_bus]))
q[t_idx] == -(b+b_to)*v[t_bus]^2 - (-b*tr+g*ti)/tm*(v[t_bus]*v[f_bus]*cos(t[f_bus]-t[t_bus])) + (-g*tr-b*ti)/tm*(v[t_bus]*v[f_bus]*sin(t[t_bus]-t[f_bus]))

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to — Method.

Do nothing, this model is symmetric

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to — Method.

nothing to do, this model is symmetric

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to — Method.

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to_on_off — Method.

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to_on_off — Method.

p[t_idx] == z*(g*v[t_bus]^2 + (-g*tr-b*ti)/tm*(v[t_bus]*v[f_bus]*cos(t[t_bus]-t[f_bus])) + (-b*tr+g*ti)/tm*(v[t_bus]*v[f_bus]*sin(t[t_bus]-t[f_bus])))
q[t_idx] == z*(-(b+c/2)*v[t_bus]^2 - (-b*tr+g*ti)/tm*(v[t_bus]*v[f_bus]*cos(t[f_bus]-t[t_bus])) + (-g*tr-b*ti)/tm*(v[t_bus]*v[f_bus]*sin(t[t_bus]-t[f_bus])))

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to_on_off — Method.

Do nothing, this model is symmetric

source

ThreePhasePowerModels.constraint_ohms_tp_yt_to_on_off — Method.

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p[t_idx] ==        g*w_to[i] + (-g*tr-b*ti)/tm*(wr[i]) + (-b*tr+g*ti)/tm*(-wi[i])
q[t_idx] == -(b+c/2)*w_to[i] - (-b*tr+g*ti)/tm*(wr[i]) + (-g*tr-b*ti)/tm*(-wi[i])

source

ThreePhasePowerModels.constraint_tp_branch_current — Method.

Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude

source

ThreePhasePowerModels.constraint_tp_branch_current — Method.

Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude

source

ThreePhasePowerModels.constraint_tp_branch_current — Method.

Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude

source

ThreePhasePowerModels.constraint_tp_branch_current — Method.

Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude

source

ThreePhasePowerModels.constraint_tp_flow_losses — Method.

Defines branch flow model power flow loss equations

source

ThreePhasePowerModels.constraint_tp_flow_losses — Method.

Defines branch flow model power flow equations

source

ThreePhasePowerModels.constraint_tp_flow_losses — Method.

Defines branch flow model power flow equations

source

ThreePhasePowerModels.constraint_tp_flow_losses — Method.

Defines branch flow model power flow equations

source

ThreePhasePowerModels.constraint_tp_storage_exchange — Method.

source

ThreePhasePowerModels.constraint_tp_theta_ref — Method.

source

ThreePhasePowerModels.constraint_tp_theta_ref — Method.

source

ThreePhasePowerModels.constraint_tp_theta_ref — Method.

nothing to do, no voltage angle variables

source

ThreePhasePowerModels.constraint_tp_theta_ref — Method.

Creates phase angle constraints at reference buses

source

ThreePhasePowerModels.constraint_tp_theta_ref — Method.

do nothing, no way to represent this in these variables

source

ThreePhasePowerModels.constraint_tp_theta_ref — Method.

source

ThreePhasePowerModels.constraint_tp_voltage — Method.

source

ThreePhasePowerModels.constraint_tp_voltage — Method.

do nothing, this model does not have complex voltage constraints

source

ThreePhasePowerModels.constraint_tp_voltage — Method.

nothing to do, these models do not have complex voltage constraints

source

ThreePhasePowerModels.constraint_tp_voltage — Method.

source

ThreePhasePowerModels.constraint_tp_voltage_magnitude_difference — Method.

Defines branch flow model power flow equations

source

ThreePhasePowerModels.constraint_tp_voltage_magnitude_difference — Method.

Defines voltage drop over a branch, linking from and to side voltage

source

ThreePhasePowerModels.constraint_tp_voltage_magnitude_difference — Method.

Defines voltage drop over a branch, linking from and to side voltage

source

ThreePhasePowerModels.constraint_tp_voltage_magnitude_difference — Method.

Defines voltage drop over a branch, linking from and to side voltage

source

ThreePhasePowerModels.constraint_voltage_magnitude_difference — Method.

source

ThreePhasePowerModels.constraint_voltage_magnitude_difference — Method.

Defines voltage drop over a branch, linking from and to side voltage magnitude

source

ThreePhasePowerModels.createCapacitor — Function.

createCapacitor(bus1, name, bus2=0; kwargs)

Creates a Dict{String,Any} containing all of the expected properties for a Capacitor. If bus2 is not specified, the capacitor will be treated as a shunt. See OpenDSS documentation for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.createGenerator — Method.

createGenerator(bus1, name; kwargs...)

Creates a Dict{String,Any} containing all of the expected properties for a Generator. See OpenDSS documentation for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.createLine — Method.

createLine(bus1, bus2, name; kwargs...)

Creates a Dict{String,Any} containing all of the properties for a Line. See OpenDSS documentation for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.createLinecode — Method.

createLinecode(name; kwargs...)

Creates a Dict{String,Any} containing all of the properties of a Linecode. See OpenDSS documentation for valid fields and ways to specify the different properties. DEPRECIATED: Calculation all done inside of createLine() due to Rg, Xg. Merge linecode values into line kwargs values BEFORE calling createLine(). This is now mainly used for parsing linecode dicts into correct data types.

source

ThreePhasePowerModels.createLoad — Method.

createLoad(bus1, name; kwargs...)

Creates a Dict{String,Any} containing all of the expected properties for a Load. See OpenDSS documentation for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.createPVSystem — Method.

createPVSystem(bus1, name; kwargs...)

Creates a Dict{String,Any} containing all of the expected properties for a PVSystem. See OpenDSS document https://github.com/tshort/OpenDSS/blob/master/Doc/OpenDSS%20PVSystem%20Model.doc for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.createReactor — Function.

createReactor(bus1, name, bus2=0; kwargs...)

Creates a Dict{String,Any} containing all of the expected properties for a Reactor. If bus2 is not specified Reactor is treated like a shunt. See OpenDSS documentation for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.createStorage — Method.

createStorage(bus1, name; kwargs...)

Creates a Dict{String,Any} containing all expected properties for a storage element. See OpenDSS documentation for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.createTransformer — Method.

createTransformer(name; kwargs...)

Creates a Dict{String,Any} containing all of the expected properties for a Transformer. See OpenDSS documentation for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.createVSource — Function.

createVSource(bus1, name, bus2=0; kwargs...)

Creates a Dict{String,Any} containing all of the expected properties for a Voltage Source. If bus2 is not specified, VSource will be treated like a generator. Mostly used as sourcebus which represents the circuit. See OpenDSS documentation for valid fields and ways to specify the different properties.

source

ThreePhasePowerModels.create_starbus — Method.

creates a starbus from a 3-winding transformer

source

ThreePhasePowerModels.discover_buses — Method.

discover_buses(dss_data)

Discovers all of the buses (not separately defined in OpenDSS), from "lines".

source

ThreePhasePowerModels.dss2tppm_branch! — Method.

dss2tppm_branch!(tppm_data, dss_data)

Adds PowerModels-style branches to tppm_data from dss_data.

source

ThreePhasePowerModels.dss2tppm_bus! — Function.

dss2tppm_bus!(tppm_data, dss_data)

Adds PowerModels-style buses to tppm_data from dss_data.

source

ThreePhasePowerModels.dss2tppm_gen! — Method.

dss2tppm_gen!(tppm_data, dss_data)

Adds PowerModels-style generators to tppm_data from dss_data.

source

ThreePhasePowerModels.dss2tppm_load! — Method.

dss2tppm_load!(tppm_data, dss_data)

Adds PowerModels-style loads to tppm_data from dss_data.

source

ThreePhasePowerModels.dss2tppm_pvsystem! — Method.

dss2tppmpvsystem!(tppmdata, dss_data)

Adds PowerModels-style pvsystems to tppm_data from dss_data.

source

ThreePhasePowerModels.dss2tppm_shunt! — Method.

dss2tppm_shunt!(tppm_data, dss_data)

Adds PowerModels-style shunts to tppm_data from dss_data.

source

ThreePhasePowerModels.dss2tppm_transformer! — Method.

dss2tppm_transformer!(tppm_data, dss_data, import_all)

Adds PowerModels-style transformers (branches) to tppm_data from dss_data.

source

ThreePhasePowerModels.find_bus — Method.

find_bus(busname, tppm_data)

Finds the index number of the bus in existing data from the given busname.

source

ThreePhasePowerModels.find_component — Method.

find_component(tppm_data, name, compType)

Returns the component of compType with name from data of type Dict{String,Array}.

source

ThreePhasePowerModels.get_dtypes — Method.

Returns a Dict{String,Type} for the desired component comp, giving all of the expected data types

source

ThreePhasePowerModels.get_linecode — Method.

returns the linecode with name id

source

ThreePhasePowerModels.get_pbs_solution — Method.

source

ThreePhasePowerModels.get_prop_name — Method.

get_prop_name(ctype, i)

Returns the ith property name for a given component type ctype.

source

ThreePhasePowerModels.get_prop_name — Method.

get_prop_name(ctype)

Returns the property names in order for a given component type ctype.

source

ThreePhasePowerModels.get_solution_tp — Method.

source

ThreePhasePowerModels.isa_array — Method.

checks if data is an opendss-style array string

source

ThreePhasePowerModels.isa_conn — Method.

checks is a string is a connection by checking the values

source

ThreePhasePowerModels.isa_matrix — Method.

checks if data is an opendss-style matrix string

source

ThreePhasePowerModels.isa_rpn — Method.

detects if expr is Reverse Polish Notation expression

source

ThreePhasePowerModels.make_mpv! — Method.

collects several fromkeys in an array and sets it to the tokey, removes from_keys

source

ThreePhasePowerModels.matlab_to_tppm — Method.

Converts a Matlab dict into a ThreePhasePowerModels dict

source

ThreePhasePowerModels.merge_dss! — Method.

merge_dss!(dss_prime, dss_to_add)

Merges two (partially) parsed OpenDSS files to the same dictionary dss_prime. Used in cases where files are referenced via the "compile" or "redirect" OpenDSS commands inside the originating file.

source

ThreePhasePowerModels.ml2pm_branch — Method.

convert raw branch data into arrays

source

ThreePhasePowerModels.ml2pm_bus — Method.

convert raw bus data into arrays

source

ThreePhasePowerModels.ml2pm_gen — Method.

convert raw generator data into arrays

source

ThreePhasePowerModels.ml2pm_load — Method.

convert raw load data into arrays

source

ThreePhasePowerModels.ml2pm_shunt — Method.

convert raw shunt data into arrays

source

ThreePhasePowerModels.objective_min_bus_power_slack — Method.

a quadratic penalty for bus power slack variables

source

ThreePhasePowerModels.parse_array — Function.

parse matrices according to active nodes

source

ThreePhasePowerModels.parse_array — Method.

parse_array(dtype, data)

Parses a OpenDSS style array string data into a one dimensional array of type dtype. Array strings are capped by either brackets, single quotes, or double quotes, and elements are separated by spaces.

source

ThreePhasePowerModels.parse_buscoords — Method.

parse_buscoords(file)

Parses a Bus Coordinate file, in either "dat" or "csv" formats, where in "dat", columns are separated by spaces, and in "csv" by commas. File expected to contain "bus,x,y" on each line.

source

ThreePhasePowerModels.parse_busname — Method.

parse_busname(busname)

Parses busnames as defined in OpenDSS, e.g. "primary.1.2.3.0".

source

ThreePhasePowerModels.parse_component — Function.

parse_component(component, properies, compDict=Dict{String,Any}())

Parses a component with properties into a compDict. If compDict is not defined, an empty dictionary will be used. Assumes that unnamed properties are given in order, but named properties can be given anywhere.

source

ThreePhasePowerModels.parse_conn — Method.

parses connection "conn" specification reducing to wye or delta

source

ThreePhasePowerModels.parse_dss — Method.

source

ThreePhasePowerModels.parse_dss — Method.

parse_dss(filename)

Parses a OpenDSS file given by filename into a Dict{Array{Dict}}. Only supports components and options, but not commands, e.g. "plot" or "solve". Will also parse files defined inside of the originating DSS file via the "compile", "redirect" or "buscoords" commands.

source

ThreePhasePowerModels.parse_dss_source_id — Method.

Parses a component's OpenDSS source information into the dss_source_id struct

source

ThreePhasePowerModels.parse_dss_with_dtypes! — Function.

parse_dss_with_dtypes!(dss_data, toParse)

Parses the data in keys defined by toParse in dss_data using types given by the default properties from the get_prop_default function.

source

ThreePhasePowerModels.parse_element_with_dtype — Method.

parses the raw dss values into their expected data types

source

ThreePhasePowerModels.parse_file — Method.

parse_file(io)

Parses the IOStream of a file into a Three-Phase PowerModels data structure.

source

ThreePhasePowerModels.parse_file — Method.

source

ThreePhasePowerModels.parse_line — Function.

parse_line(elements, curCompDict=Dict{String,Any}())

Parses an already separated line given by elements (an array) of an OpenDSS file into curCompDict. If not defined, curCompDict is an empty dictionary.

source

ThreePhasePowerModels.parse_matlab — Method.

source

ThreePhasePowerModels.parse_matlab — Method.

source

ThreePhasePowerModels.parse_matlab_string — Method.

source

ThreePhasePowerModels.parse_matrix — Method.

parse_matrix(dtype, data)

Parses a OpenDSS style triangular matrix string data into a two dimensional array of type dtype. Matrix strings are capped by either parenthesis or brackets, rows are separated by "|", and columns are separated by spaces.

source

ThreePhasePowerModels.parse_matrix — Method.

parse matrices according to active nodes

source

ThreePhasePowerModels.parse_opendss — Method.

Parses a Dict resulting from the parsing of a DSS file into a PowerModels usable format.

source

ThreePhasePowerModels.parse_opendss — Method.

Parses a DSS file into a PowerModels usable format.

source

ThreePhasePowerModels.parse_options — Method.

parse_options(options)

Parses options defined with the set command in OpenDSS.

source

ThreePhasePowerModels.parse_properties — Method.

parse_properties(properties)

Parses a string of properties of a component type, character by character into an array with each element containing (if present) the property name, "=", and the property value.

source

ThreePhasePowerModels.parse_rpn — Function.

parses Reverse Polish Notation expr

source

ThreePhasePowerModels.post_mn_tp_strg_opf — Method.

source

ThreePhasePowerModels.post_tp_opf — Method.

source

ThreePhasePowerModels.post_tp_opf_bf — Method.

source

ThreePhasePowerModels.post_tp_opf_pbs — Method.

source

ThreePhasePowerModels.post_tp_ots — Method.

source

ThreePhasePowerModels.post_tp_pf — Method.

source

ThreePhasePowerModels.post_tp_pf_bf — Method.

source

ThreePhasePowerModels.post_tp_pf_pbs — Method.

source

ThreePhasePowerModels.post_tp_strg_opf — Method.

source

ThreePhasePowerModels.relaxation_psd_to_psd_real — Method.

real-valued SDP to SDP relaxation based on PSDness of principal minors, default is 3x3 SDP relaxation

source

ThreePhasePowerModels.relaxation_psd_to_soc — Method.

See section 4.3 for complex to real PSD constraint transformation: @article{Fazel2001, author = {Fazel, M. and Hindi, H. and Boyd, S.P.}, title = {{A rank minimization heuristic with application to minimum order system approximation}}, doi = {10.1109/ACC.2001.945730}, journal = {Proc. American Control Conf.}, number = {2}, pages = {4734–4739}, url = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=945730}, volume = {6}, year = {2001} }

source

ThreePhasePowerModels.relaxation_psd_to_soc_complex — Method.

SDP to SOC relaxation of type 2, applied to complex-value matrix, as described in:

@article{Kim2003,
author = {Kim, S and Kojima, M and Yamashita, M},
title = {{Second order cone programming relaxation of a positive semidefinite constraint}},
doi = {10.1080/1055678031000148696},
journal = {Optimization Methods and Software},
number = {5},
pages = {535--541},
volume = {18},
year = {2003}
}

source

ThreePhasePowerModels.relaxation_psd_to_soc_complex_conic — Method.

SDP to SOC relaxation of type 2, applied to complex-value matrix, as described in:

@article{Kim2003,
author = {Kim, S and Kojima, M and Yamashita, M},
title = {{Second order cone programming relaxation of a positive semidefinite constraint}},
doi = {10.1080/1055678031000148696},
journal = {Optimization Methods and Software},
number = {5},
pages = {535--541},
volume = {18},
year = {2003}
}

source

ThreePhasePowerModels.relaxation_psd_to_soc_conic — Method.

See section 4.3 for complex to real PSD constraint transformation: @article{Fazel2001, author = {Fazel, M. and Hindi, H. and Boyd, S.P.}, title = {{A rank minimization heuristic with application to minimum order system approximation}}, doi = {10.1109/ACC.2001.945730}, journal = {Proc. American Control Conf.}, number = {2}, pages = {4734–4739}, url = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=945730}, volume = {6}, year = {2001} }

source

ThreePhasePowerModels.relaxation_psd_to_soc_real — Method.

SDP to SOC relaxation of type 2, applied to real-value matrix, as described in:

@article{Kim2003,
author = {Kim, S and Kojima, M and Yamashita, M},
title = {{Second order cone programming relaxation of a positive semidefinite constraint}},
doi = {10.1080/1055678031000148696},
journal = {Optimization Methods and Software},
number = {5},
pages = {535--541},
volume = {18},
year = {2003}
}

source

ThreePhasePowerModels.relaxation_psd_to_soc_real_conic — Method.

SDP to SOC relaxation of type 2, applied to real-value matrix, as described in:

@article{Kim2003,
author = {Kim, S and Kojima, M and Yamashita, M},
title = {{Second order cone programming relaxation of a positive semidefinite constraint}},
doi = {10.1080/1055678031000148696},
journal = {Optimization Methods and Software},
number = {5},
pages = {535--541},
volume = {18},
year = {2003}
}

source

ThreePhasePowerModels.roll — Method.

rolls a 1d array left or right by idx

source

ThreePhasePowerModels.run_dc_tp_pf — Method.

source

ThreePhasePowerModels.run_mn_tp_strg_opf — Method.

multi-network opf with storage

source

ThreePhasePowerModels.run_tp_strg_opf — Method.

opf with storage

source

ThreePhasePowerModels.set_default — Method.

checks if the given dict has a value, if not, sets a default value

source

ThreePhasePowerModels.sqr — Method.

Squares x, for parsing Reverse Polish Notation

source

ThreePhasePowerModels.strip_comments — Method.

Strips comments, defined by "!" from the ends of lines

source

ThreePhasePowerModels.strip_lines — Method.

strips lines that are either commented (block or single) or empty

source

ThreePhasePowerModels.to_sym_keys — Method.

converts Dict{String,Any} to Dict{Symbol,Any} for passing as kwargs

source

ThreePhasePowerModels.translate_version! — Method.

Translates legacy versions into current version format

source

ThreePhasePowerModels.variable_active_bus_power_slack — Method.

source

ThreePhasePowerModels.variable_bus_power_slack — Method.

generates variables for both active and reactive slack at each bus

source

ThreePhasePowerModels.variable_lower_triangle_active_branch_flow — Method.

variable: p_lt[l,i,j] for (l,i,j) in arcs

source

ThreePhasePowerModels.variable_lower_triangle_reactive_branch_flow — Method.

variable: q_lt[l,i,j] for (l,i,j) in arcs

source

ThreePhasePowerModels.variable_reactive_bus_power_slack — Method.

source

ThreePhasePowerModels.variable_tp_branch_flow — Method.

source

ThreePhasePowerModels.variable_tp_branch_flow — Method.

source

ThreePhasePowerModels.variable_tp_branch_flow — Method.

source

ThreePhasePowerModels.variable_tp_branch_flow — Method.

source

ThreePhasePowerModels.variable_tp_storage — Method.

variables for modeling storage units, includes grid injection and internal variables

source

ThreePhasePowerModels.variable_tp_voltage — Method.

source

ThreePhasePowerModels.variable_tp_voltage — Method.

source

ThreePhasePowerModels.variable_tp_voltage — Method.

source

ThreePhasePowerModels.variable_tp_voltage_magnitude_sqr — Method.

variable: w[i] >= 0 for i in buses

source

ThreePhasePowerModels.variable_tp_voltage_prod_hermitian — Method.

source

ThreePhasePowerModels.variable_tp_voltage_prod_hermitian — Method.

source

ThreePhasePowerModels.variable_tp_voltage_product — Method.

source

ThreePhasePowerModels.variable_upper_triangle_active_branch_flow — Method.

variable: p_ut[l,i,j] for (l,i,j) in arcs

source

ThreePhasePowerModels.variable_upper_triangle_reactive_branch_flow — Method.

variable: q_ut[l,i,j] for (l,i,j) in arcs

source

ThreePhasePowerModels.where_is_comp — Method.

where_is_comp(data, comp_id)

Finds existing component of id comp_id in array of data and returns index. Assumes all components in data are unique.

source