Constraints

PowerModelsDistribution.constraint_M_psd — Method

For rectangular coordinates of a complex matrix M=Mre+im*Mim, this function applies constraints equivalent to requiring that M itself is PSD.

PowerModelsDistribution.constraint_SWL_psd — Method

Take a multi-conductor voltage variable V and a current variable I. The associated power is then defined as S = VI^H Define the lifted variables as W and L as W = VV^H, L = I*I^H Then, it is equally valid that [W S; S^H L] ∈ PSDCone, rank([W S; S^H L])=1 This function adds this PSD constraint for the rectangular coordinates of S, W and L.

PowerModelsDistribution.constraint_mc_branch_current_series_product_hermitian — Method

PowerModelsDistribution.constraint_mc_bus_voltage_balance — Method

Impose all balance related constraints for which key present in data model of bus. For a discussion of sequence components and voltage unbalance factor (VUF), see @INPROCEEDINGS{girigoudarmolzahnroald-2019, author={K. Girigoudar and D. K. Molzahn and L. A. Roald}, booktitle={submitted}, title={{Analytical and Empirical Comparisons of Voltage Unbalance Definitions}}, year={2019}, month={}, url={https://molzahn.github.io/pubs/girigoudarmolzahnroald-2019.pdf} }

PowerModelsDistribution.constraint_mc_bus_voltage_drop — Method

Defines voltage drop over a branch, linking from and to side complex voltage

PowerModelsDistribution.constraint_mc_bus_voltage_drop — Method

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_ll — Method

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_negative_sequence — Method

a = exp(im2π/3) U+ = (1Ua + aUb a^2Uc)/3 U- = (1Ua + a^2Ub aUc)/3 vuf = |U-|/|U+| |U-| <= vufmax|U+| |U-|^2 <= vufmax^2*|U+|^2

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_on_off — Method

on/off bus voltage magnitude constraint

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_on_off — Method

on/off bus voltage magnitude constraint

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_on_off — Method

on/off voltage magnitude constraint

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_positive_sequence — Method

a = exp(im2π/3) U+ = (1Ua + aUb a^2Uc)/3 U- = (1Ua + a^2Ub aUc)/3 vuf = |U-|/|U+| |U-| <= vufmax|U+| |U-|^2 <= vufmax^2*|U+|^2

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_sqr_on_off — Method

on/off bus voltage magnitude squared constraint for relaxed formulations

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_sqr_on_off — Method

on/off voltage magnitude squared constraint for relaxed formulations

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_vuf — Method

a = exp(im2π/3) U+ = (1Ua + aUb a^2Uc)/3 U- = (1Ua + a^2Ub aUc)/3 vuf = |U-|/|U+| |U-| <= vufmax|U+| |U-|^2 <= vufmax^2*|U+|^2

PowerModelsDistribution.constraint_mc_bus_voltage_magnitude_zero_sequence — Method

a = exp(im2π/3) U+ = (1Ua + aUb a^2Uc)/3 U- = (1Ua + a^2Ub aUc)/3 vuf = |U-|/|U+| |U-| <= vufmax|U+| |U-|^2 <= vufmax^2*|U+|^2

PowerModelsDistribution.constraint_mc_bus_voltage_on_off — Method

on/off bus voltage constraint for relaxed forms

PowerModelsDistribution.constraint_mc_bus_voltage_on_off — Method

bus voltage on/off constraint for load shed problem

PowerModelsDistribution.constraint_mc_bus_voltage_on_off — Method

bus voltage on/off constraint for load shed problem

PowerModelsDistribution.constraint_mc_bus_voltage_on_off — Method

on/off bus voltage constraint for DCP formulation, nothing to do

PowerModelsDistribution.constraint_mc_bus_voltage_on_off — Method

on/off constraint for bus voltages

PowerModelsDistribution.constraint_mc_current_balance — Method

Kirchhoff's current law applied to buses sum(cr + im*ci) = 0

PowerModelsDistribution.constraint_mc_current_balance — Method

PowerModelsDistribution.constraint_mc_current_from — Method

Defines how current distributes over series and shunt impedances of a pi-model branch

PowerModelsDistribution.constraint_mc_current_from — Method

PowerModelsDistribution.constraint_mc_current_limit — Method

PowerModelsDistribution.constraint_mc_current_limit — Method

Bounds the current magnitude at both from and to side of a branch cr[f_idx]^2 + ci[f_idx]^2 <= c_rating_a^2 cr[t_idx]^2 + ci[t_idx]^2 <= c_rating_a^2

PowerModelsDistribution.constraint_mc_current_to — Method

Defines how current distributes over series and shunt impedances of a pi-model branch

PowerModelsDistribution.constraint_mc_current_to — Method

PowerModelsDistribution.constraint_mc_gen_active_bounds — Method

pmin <= Re(v*cg') <= pmax

PowerModelsDistribution.constraint_mc_gen_active_bounds — Method

defines limits on active power output of a generator where bounds can't be used

PowerModelsDistribution.constraint_mc_gen_power_on_off — Method

on/off constraint for generators

PowerModelsDistribution.constraint_mc_gen_power_on_off — Method

on/off constraint for generators

PowerModelsDistribution.constraint_mc_gen_power_on_off — Method

PowerModelsDistribution.constraint_mc_gen_power_setpoint_real — Method

pg[i] == pg

PowerModelsDistribution.constraint_mc_gen_power_setpoint_real — Method

PowerModelsDistribution.constraint_mc_gen_power_setpoint_real — Method

generator active power setpoint constraint

PowerModelsDistribution.constraint_mc_gen_reactive_bounds — Method

qmin <= Im(v*cg') <= qmax

PowerModelsDistribution.constraint_mc_gen_reactive_bounds — Method

defines limits on reactive power output of a generator where bounds can't be used

PowerModelsDistribution.constraint_mc_generator_power — Method

Only support wye-connected generators.

PowerModelsDistribution.constraint_mc_generator_power — Method

DELTA When connected in delta, the load power gives the reference in the delta reference frame. This means sd1 = vab.conj(iab) = (va-vb).conj(iab) We can relate this to the per-phase power by sna = va.conj(ia) = va.conj(iab-ica) = va.conj(conj(sab/vab) - conj(sca/vca)) = va.(sab/(va-vb) - sca/(vc-va)) So for delta, sn is constrained indirectly.

PowerModelsDistribution.constraint_mc_generator_power — Method

Link the current and power withdrawn by a generator at the bus through a PSD constraint. The rank-1 constraint is dropped in this formulation.

PowerModelsDistribution.constraint_mc_generator_power_delta — Method

PowerModelsDistribution.constraint_mc_generator_power_delta — Method

PowerModelsDistribution.constraint_mc_generator_power_delta — Method

delta connected generator setpoint constraint for IVR formulation

PowerModelsDistribution.constraint_mc_generator_power_wye — Method

PowerModelsDistribution.constraint_mc_generator_power_wye — Method

wye connected generator setpoint constraint for IVR formulation

PowerModelsDistribution.constraint_mc_load_current_delta — Method

We want to express sab = cp.|vab|+im.cq.|vab| iab = conj(sab/vab) = |vab|.(cq-im.cq)/conj(vab) = (1/|vab|).(cp-im.cq)*vab idem for ibc and ica And then sa = va.conj(ia) = va.conj(iab-ica) idem for sb and sc

PowerModelsDistribution.constraint_mc_load_power — Method

Creates the constraints modelling the (relaxed) voltage-dependent loads.

PowerModelsDistribution.constraint_mc_load_power — Method

Only support wye-connected, constant-power loads.

PowerModelsDistribution.constraint_mc_load_power — Method

CONSTANT POWER Fixes the load power sd. sd = [sd1, sd2, sd3] What is actually fixed, depends on whether the load is connected in delta or wye. When connected in wye, the load power equals the per-phase power sn drawn at the bus to which the load is connected. sd1 = va.conj(ia) = sn_a

CONSTANT CURRENT Sets the active and reactive load power sd to be proportional to the the voltage magnitude. pd = cp.|vm| qd = cq.|vm| sd = cp.|vm| + j.cq.|vm|

CONSTANT IMPEDANCE Sets the active and reactive power drawn by the load to be proportional to the square of the voltage magnitude. pd = cp.|vm|^2 qd = cq.|vm|^2 sd = cp.|vm|^2 + j.cq.|vm|^2

DELTA When connected in delta, the load power gives the reference in the delta reference frame. This means sd1 = vab.conj(iab) = (va-vb).conj(iab) We can relate this to the per-phase power by sna = va.conj(ia) = va.conj(iab-ica) = va.conj(conj(sab/vab) - conj(sca/vca)) = va.(sab/(va-vb) - sca/(vc-va)) So for delta, sn is constrained indirectly.

PowerModelsDistribution.constraint_mc_load_power — Method

Creates the constraints modelling the (relaxed) voltage-dependent loads for the matrix KCL formulation.

PowerModelsDistribution.constraint_mc_load_power_delta — Method

PowerModelsDistribution.constraint_mc_load_power_delta — Method

PowerModelsDistribution.constraint_mc_load_power_delta — Method

delta connected load setpoint constraint for IVR formulation

PowerModelsDistribution.constraint_mc_load_power_wye — Method

PowerModelsDistribution.constraint_mc_load_power_wye — Method

PowerModelsDistribution.constraint_mc_load_power_wye — Method

wye connected load setpoint constraint for IVR formulation

PowerModelsDistribution.constraint_mc_model_current — Method

Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude

PowerModelsDistribution.constraint_mc_model_current — Method

PowerModelsDistribution.constraint_mc_model_current — Method

Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude

PowerModelsDistribution.constraint_mc_model_current — Method

Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude

PowerModelsDistribution.constraint_mc_model_current — Method

Defines relationship between branch (series) power flow, branch (series) current and node voltage magnitude

PowerModelsDistribution.constraint_mc_model_voltage — Method

nothing to do, these models do not have complex voltage constraints

PowerModelsDistribution.constraint_mc_model_voltage — Method

do nothing by default

PowerModelsDistribution.constraint_mc_model_voltage — Method

PowerModelsDistribution.constraint_mc_model_voltage_magnitude_difference — Method

Defines voltage drop over a branch, linking from and to side voltage

PowerModelsDistribution.constraint_mc_model_voltage_magnitude_difference — Method

PowerModelsDistribution.constraint_mc_model_voltage_magnitude_difference — Method

Defines voltage drop over a branch, linking from and to side voltage

PowerModelsDistribution.constraint_mc_network_power_balance — Method

PowerModelsDistribution.constraint_mc_network_power_balance — Method

ensures that power generation and demand are balanced

PowerModelsDistribution.constraint_mc_ohms_yt_from — Method

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

PowerModelsDistribution.constraint_mc_ohms_yt_from — Method

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p_fr ==     g[c,c] * vm_fr[c]^2 +
            sum( g[c,d]*vm_fr[c]*vm_fr[d]*cos(va_fr[c]-va_fr[d]) +
                 b[c,d]*vm_fr[c]*vm_fr[d]*sin(va_fr[c]-va_fr[d]) for d in conductor_ids(pm) if d != c) +
            sum(-g[c,d]*vm_fr[c]*vm_to[d]*cos(va_fr[c]-va_to[d]) +
                -b[c,d]*vm_fr[c]*vm_to[d]*sin(va_fr[c]-va_to[d]) for d in conductor_ids(pm))
            + g_fr[c,c] * vm_fr[c]^2 +
            sum( g_fr[c,d]*vm_fr[c]*vm_fr[d]*cos(va_fr[c]-va_fr[d]) +
                 b_fr[c,d]*vm_fr[c]*vm_fr[d]*sin(va_fr[c]-va_fr[d]) for d in conductor_ids(pm) if d != c)
            )
q_fr == -b[c,c] *vm_fr[c]^2 -
            sum( b[c,d]*vm_fr[c]*vm_fr[d]*cos(va_fr[c]-va_fr[d]) -
                 g[c,d]*vm_fr[c]*vm_fr[d]*sin(va_fr[c]-va_fr[d]) for d in conductor_ids(pm) if d != c) -
            sum(-b[c,d]*vm_fr[c]*vm_to[d]*cos(va_fr[c]-va_to[d]) +
                 g[c,d]*vm_fr[c]*vm_to[d]*sin(va_fr[c]-va_to[d]) for d in conductor_ids(pm))
            -b_fr[c,c] *vm_fr[c]^2 -
            sum( b_fr[c,d]*vm_fr[c]*vm_fr[d]*cos(va_fr[c]-va_fr[d]) -
                 g_fr[c,d]*vm_fr[c]*vm_fr[d]*sin(va_fr[c]-va_fr[d]) for d in conductor_ids(pm) if d != c)
            )

PowerModelsDistribution.constraint_mc_ohms_yt_from — Method

Creates Ohms constraints

sfr = vfr.conj(Y(vfr-vto)) sfr = (vrfr+imvi_fr).(G-imB)([vrfr-vrto]-im[vifr-vito]) sfr = (vrfr+imvifr).([Gvrfr-Gvr_to-Bvifr+B*vito]-im[Gvifr-G*vito+Bvr_fr-Bvr_to])

PowerModelsDistribution.constraint_mc_ohms_yt_from — Method

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p[f_idx] == -b*(t[f_bus] - t[t_bus])

PowerModelsDistribution.constraint_mc_ohms_yt_from — Method

ohms constraint for branches on the from-side

PowerModelsDistribution.constraint_mc_ohms_yt_from — Method

nothing to do, no voltage angle variables

PowerModelsDistribution.constraint_mc_ohms_yt_to — Method

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

PowerModelsDistribution.constraint_mc_ohms_yt_to — Method

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p[t_idx] ==  (g+g_to)*v[t_bus]^2 + (-g*tr-b*ti)/tm*(v[t_bus]*v[f_bus]*cos(t[t_bus]-t[f_bus])) + (-b*tr+g*ti)/tm*(v[t_bus]*v[f_bus]*sin(t[t_bus]-t[f_bus]))
q[t_idx] == -(b+b_to)*v[t_bus]^2 - (-b*tr+g*ti)/tm*(v[t_bus]*v[f_bus]*cos(t[f_bus]-t[t_bus])) + (-g*tr-b*ti)/tm*(v[t_bus]*v[f_bus]*sin(t[t_bus]-t[f_bus]))

PowerModelsDistribution.constraint_mc_ohms_yt_to — Method

Creates Ohms constraints (yt post fix indicates that Y and T values are in rectangular form)

p[t_idx] ==  (g+g_to)*v[t_bus]^2 + (-g*tr-b*ti)/tm*(v[t_bus]*v[f_bus]*cos(t[t_bus]-t[f_bus])) + (-b*tr+g*ti)/tm*(v[t_bus]*v[f_bus]*sin(t[t_bus]-t[f_bus]))
q[t_idx] == -(b+b_to)*v[t_bus]^2 - (-b*tr+g*ti)/tm*(v[t_bus]*v[f_bus]*cos(t[f_bus]-t[t_bus])) + (-g*tr-b*ti)/tm*(v[t_bus]*v[f_bus]*sin(t[t_bus]-t[f_bus]))

PowerModelsDistribution.constraint_mc_ohms_yt_to — Method

Do nothing, this model is symmetric

PowerModelsDistribution.constraint_mc_ohms_yt_to — Method

ohms constraint for branches on the to-side

PowerModelsDistribution.constraint_mc_ohms_yt_to — Method

nothing to do, this model is symmetric

PowerModelsDistribution.constraint_mc_oltc_tap_fix — Method

For a variable tap transformer, fix the tap variables which are fixed. For example, an OLTC where the third phase is fixed, will have tap variables for all phases, but the third tap variable should be fixed.

PowerModelsDistribution.constraint_mc_power_balance — Method

PowerModelsDistribution.constraint_mc_power_balance — Method

PowerModelsDistribution.constraint_mc_power_balance — Method

PowerModelsDistribution.constraint_mc_power_balance — Method

power balanace constraint with line shunts and transformers, active power only

PowerModelsDistribution.constraint_mc_power_balance — Method

KCL including transformer arcs and load variables.

PowerModelsDistribution.constraint_mc_power_balance — Method

PowerModelsDistribution.constraint_mc_power_balance — Method

Shunt handling in matrix form: I = Y.U S = U.I' = U.(Y.U)' = U.U'.Y' = W.Y' = (Wr+j.Wi)(G+jB)' = (Wr+j.Wi)(G'-j.B') = (Wr.G'+Wi.B')+j(-Wr.B'+Wi.G') P = Wr.G'+Wi.B' Q = -Wr.B'+Wi.G'

PowerModelsDistribution.constraint_mc_power_balance_shed — Method

KCL for load shed problem with transformers (AbstractWForms)

PowerModelsDistribution.constraint_mc_power_balance_shed — Method

PowerModelsDistribution.constraint_mc_power_balance_shed — Method

PowerModelsDistribution.constraint_mc_power_balance_shed — Method

power balance constraint with line shunts and transformers for load shed problem, DCP formulation

PowerModelsDistribution.constraint_mc_power_balance_shed — Method

KCL for load shed problem with transformers

PowerModelsDistribution.constraint_mc_power_balance_shed_simple — Method

power balance constraint with line shunts and transformers for load shed problem, DCP formulation

PowerModelsDistribution.constraint_mc_power_balance_simple — Method

PowerModelsDistribution.constraint_mc_power_balance_simple — Method

PowerModelsDistribution.constraint_mc_power_balance_simple — Method

KCL including transformer arcs

PowerModelsDistribution.constraint_mc_power_balance_slack — Method

PowerModelsDistribution.constraint_mc_power_balance_slack — Method

PowerModelsDistribution.constraint_mc_power_balance_slack — Method

PowerModelsDistribution.constraint_mc_power_balance_slack — Method

PowerModelsDistribution.constraint_mc_power_losses — Method

Defines branch flow model power flow equations

PowerModelsDistribution.constraint_mc_power_losses — Method

PowerModelsDistribution.constraint_mc_power_losses — Method

Defines branch flow model power flow equations

PowerModelsDistribution.constraint_mc_regen_setpoint_active — Method

qq[i] == qq

PowerModelsDistribution.constraint_mc_storage_current_limit — Method

PowerModelsDistribution.constraint_mc_storage_current_limit — Method

PowerModelsDistribution.constraint_mc_storage_current_limit — Method

PowerModelsDistribution.constraint_mc_storage_losses — Method

Neglects the active and reactive loss terms associated with the squared current magnitude.

PowerModelsDistribution.constraint_mc_storage_losses — Method

PowerModelsDistribution.constraint_mc_storage_losses — Method

PowerModelsDistribution.constraint_mc_storage_losses — Method

storage loss constraints, delegate to PowerModels

PowerModelsDistribution.constraint_mc_storage_on_off — Method

PowerModelsDistribution.constraint_mc_storage_on_off — Method

PowerModelsDistribution.constraint_mc_storage_on_off — Method

PowerModelsDistribution.constraint_mc_storage_thermal_limit — Method

PowerModelsDistribution.constraint_mc_storage_thermal_limit — Method

PowerModelsDistribution.constraint_mc_storage_thermal_limit — Method

PowerModelsDistribution.constraint_mc_switch_current_limit — Method

PowerModelsDistribution.constraint_mc_switch_power_on_off — Method

PowerModelsDistribution.constraint_mc_switch_power_on_off — Method

PowerModelsDistribution.constraint_mc_switch_state — Method

PowerModelsDistribution.constraint_mc_switch_state_closed — Method

PowerModelsDistribution.constraint_mc_switch_state_closed — Method

PowerModelsDistribution.constraint_mc_switch_state_closed — Method

PowerModelsDistribution.constraint_mc_switch_state_closed — Method

PowerModelsDistribution.constraint_mc_switch_state_closed — Method

PowerModelsDistribution.constraint_mc_switch_state_closed — Method

nothing to do

PowerModelsDistribution.constraint_mc_switch_state_on_off — Method

PowerModelsDistribution.constraint_mc_switch_state_on_off — Method

PowerModelsDistribution.constraint_mc_switch_state_on_off — Method

PowerModelsDistribution.constraint_mc_switch_state_on_off — Method

PowerModelsDistribution.constraint_mc_switch_state_on_off — Method

PowerModelsDistribution.constraint_mc_switch_state_open — Method

PowerModelsDistribution.constraint_mc_switch_state_open — Method

PowerModelsDistribution.constraint_mc_switch_thermal_limit — Method

PowerModelsDistribution.constraint_mc_thermal_limit_from — Method

-rate_a <= p[f_idx] <= rate_a

PowerModelsDistribution.constraint_mc_thermal_limit_from — Method

p[f_idx]^2 + q[f_idx]^2 <= rate_a^2

PowerModelsDistribution.constraint_mc_thermal_limit_from — Method

Generic thermal limit constraint from-side

PowerModelsDistribution.constraint_mc_thermal_limit_from — Method

branch thermal constraints from

PowerModelsDistribution.constraint_mc_thermal_limit_from_ne — Method

PowerModelsDistribution.constraint_mc_thermal_limit_from_on_off — Method

PowerModelsDistribution.constraint_mc_thermal_limit_to — Method

PowerModelsDistribution.constraint_mc_thermal_limit_to — Method

p[t_idx]^2 + q[t_idx]^2 <= rate_a^2

PowerModelsDistribution.constraint_mc_thermal_limit_to — Method

Generic thermal limit constraint to-side

PowerModelsDistribution.constraint_mc_thermal_limit_to — Method

branch thermal constraints to

PowerModelsDistribution.constraint_mc_thermal_limit_to_ne — Method

PowerModelsDistribution.constraint_mc_thermal_limit_to_on_off — Method

PowerModelsDistribution.constraint_mc_theta_ref — Method

do nothing, no way to represent this in these variables

PowerModelsDistribution.constraint_mc_theta_ref — Method

PowerModelsDistribution.constraint_mc_theta_ref — Method

Creates phase angle constraints at reference buses

PowerModelsDistribution.constraint_mc_theta_ref — Method

Creates phase angle constraints at reference buses

PowerModelsDistribution.constraint_mc_theta_ref — Method

reference angle constraints

PowerModelsDistribution.constraint_mc_theta_ref — Method

balanced three-phase phasor

PowerModelsDistribution.constraint_mc_theta_ref — Method

nothing to do, no voltage angle variables

PowerModelsDistribution.constraint_mc_transformer_power — Method

Transformer constraints, considering winding type, conductor order, polarity and tap settings.

PowerModelsDistribution.constraint_mc_transformer_power — Method

nothing to do, this model is symmetric

PowerModelsDistribution.constraint_mc_transformer_power_dy — Method

PowerModelsDistribution.constraint_mc_transformer_power_dy — Method

This function adds all constraints required to model a two-winding, delta-wye connected transformer.

PowerModelsDistribution.constraint_mc_transformer_power_dy — Method

delta-wye transformer power constraint for IVR formulation

PowerModelsDistribution.constraint_mc_transformer_power_yy — Method

PowerModelsDistribution.constraint_mc_transformer_power_yy — Method

This function adds all constraints required to model a two-winding, wye-wye connected transformer.

PowerModelsDistribution.constraint_mc_transformer_power_yy — Method

wye-wye transformer power constraint for IVR formulation

PowerModelsDistribution.constraint_mc_transformer_power_yy — Method

Links to and from power and voltages in a wye-wye transformer, assumes tmfixed is true wfri=(politm_scaletmi)^2wto_i

PowerModelsDistribution.constraint_mc_voltage_angle_difference — Method

This is duplicated at PowerModelsDistribution level to correctly handle the indexing of the shunts.

PowerModelsDistribution.constraint_mc_voltage_angle_difference — Method

PowerModelsDistribution.constraint_mc_voltage_angle_difference — Method

Bounds the voltage angle difference between bus pairs

PowerModelsDistribution.constraint_mc_voltage_angle_difference — Method

nothing to do, these models do not have angle difference constraints

PowerModelsDistribution.constraint_mc_voltage_angle_difference — Method

PowerModelsDistribution.constraint_mc_voltage_angle_difference — Method

This is duplicated at PowerModelsDistribution level to correctly handle the indexing of the shunts.

PowerModelsDistribution.constraint_mc_voltage_magnitude_bounds — Method

vmin <= vm[i] <= vmax

PowerModelsDistribution.constraint_mc_voltage_magnitude_bounds — Method

This constraint captures problem agnostic constraints that define limits for voltage magnitudes (where variable bounds cannot be used) Notable examples include IVRUPowerModel and ACRUPowerModel

PowerModelsDistribution.constraint_mc_voltage_magnitude_only — Method

vm[i] == vmref

PowerModelsDistribution.constraint_mc_voltage_magnitude_only — Method

vm[i] == vmref

PowerModelsDistribution.constraint_mc_voltage_magnitude_only — Method

vm[i] == vmref

PowerModelsDistribution.constraint_mc_voltage_magnitude_only — Method

nothing to do

PowerModelsDistribution.constraint_mc_voltage_magnitude_only — Method

voltage magnitude setpoint constraint

PowerModelsDistribution.constraint_mc_voltage_psd — Method

Add explicit PSD-ness of W for nodes where it is not implied

PowerModelsDistribution.constraint_mc_voltage_psd — Method

Add explicit PSD-ness of W for nodes where it is not implied

PowerModelsDistribution.constraint_mc_voltage_psd — Method

Add explicit PSD-ness of W for nodes where it is not implied

PowerModelsDistribution.constraint_pqw — Method

Creates the constraints modelling the (relaxed) voltage-dependency of the power consumed in each phase, s=p+jq. This is completely symmetrical for p and q, with appropriate substitutions of the variables and parameters: p->q, a->b, alpha->beta, pmin->qmin, pmax->qmax

PowerModelsDistribution.constraint_storage_complementarity_mi — Method

PowerModelsDistribution.constraint_storage_complementarity_mi — Method

PowerModelsDistribution.constraint_storage_complementarity_nl — Method

PowerModelsDistribution.constraint_storage_complementarity_nl — Method

PowerModelsDistribution.constraint_storage_losses — Method

Neglects the active and reactive loss terms associated with the squared current magnitude.

PowerModelsDistribution.constraint_storage_losses — Method

PowerModelsDistribution.constraint_storage_losses — Method

PowerModelsDistribution.constraint_storage_losses — Method

PowerModelsDistribution.constraint_storage_losses — Method

PowerModelsDistribution.constraint_storage_losses — Method

PowerModelsDistribution.constraint_storage_state — Method

PowerModelsDistribution.constraint_storage_state — Method

PowerModelsDistribution.constraint_storage_state — Method

PowerModelsDistribution.constraint_storage_state_initial — Method

PowerModelsDistribution.constraint_switch_thermal_limit — Method

Relaxation Helpers

PowerModelsDistribution.cut_complex_product_and_angle_difference — Method

cut_complex_product_and_angle_difference(m, wf, wt, wr, wi, angmin, angmax)

A valid inequality for the product of two complex variables with magnitude and angle difference bounds. In the literature this constraints are called the Lifted Nonlinear Cuts (LNCs). @misc{1512.04644, Author = {Carleton Coffrin and Hassan Hijazi and Pascal Van Hentenryck}, Title = {Strengthening the SDP Relaxation of AC Power Flows with Convex Envelopes, Bound Tightening, and Lifted Nonlinear Cuts}, Year = {2015}, Eprint = {arXiv:1512.04644}, }

PowerModelsDistribution.relaxation_psd_to_psd_real — Method

relaxation_psd_to_psd_real(m, mxreal, mximag; ndim=3)

For debugging / exploration: real-valued SDP to SDP relaxation based on PSDness of principal minors, default is 3x3 SDP relaxation

PowerModelsDistribution.relaxation_psd_to_soc — Method

relaxation_psd_to_soc(m::JuMP.Model, mxreal, mximag; complex::Bool=true)

See section 4.3 for complex to real PSD constraint transformation: @article{Fazel2001, author = {Fazel, M. and Hindi, H. and Boyd, S.P.}, title = {{A rank minimization heuristic with application to minimum order system approximation}}, doi = {10.1109/ACC.2001.945730}, journal = {Proc. American Control Conf.}, number = {2}, pages = {4734–4739}, url = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=945730}, volume = {6}, year = {2001} }

PowerModelsDistribution.relaxation_psd_to_soc_complex — Method

relaxation_psd_to_soc_complex(m, mxreal, mximag)

SDP to SOC relaxation of type 2, applied to complex-value matrix, as described in:

@article{Kim2003,
author = {Kim, S and Kojima, M and Yamashita, M},
title = {{Second order cone programming relaxation of a positive semidefinite constraint}},
doi = {10.1080/1055678031000148696},
journal = {Optimization Methods and Software},
number = {5},
pages = {535--541},
volume = {18},
year = {2003}
}

PowerModelsDistribution.relaxation_psd_to_soc_complex_conic — Method

relaxation_psd_to_soc_complex_conic(m, mxreal, mximag)

SDP to SOC relaxation of type 2, applied to complex-value matrix, as described in:

@article{Kim2003,
author = {Kim, S and Kojima, M and Yamashita, M},
title = {{Second order cone programming relaxation of a positive semidefinite constraint}},
doi = {10.1080/1055678031000148696},
journal = {Optimization Methods and Software},
number = {5},
pages = {535--541},
volume = {18},
year = {2003}
}

PowerModelsDistribution.relaxation_psd_to_soc_conic — Method

relaxation_psd_to_soc_conic(m, mxreal, mximag; complex=true)

See section 4.3 for complex to real PSD constraint transformation: @article{Fazel2001, author = {Fazel, M. and Hindi, H. and Boyd, S.P.}, title = {{A rank minimization heuristic with application to minimum order system approximation}}, doi = {10.1109/ACC.2001.945730}, journal = {Proc. American Control Conf.}, number = {2}, pages = {4734–4739}, url = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=945730}, volume = {6}, year = {2001} }

PowerModelsDistribution.relaxation_psd_to_soc_real — Method

relaxation_psd_to_soc_real(m, mx)

SDP to SOC relaxation of type 2, applied to real-value matrix, as described in:

@article{Kim2003,
author = {Kim, S and Kojima, M and Yamashita, M},
title = {{Second order cone programming relaxation of a positive semidefinite constraint}},
doi = {10.1080/1055678031000148696},
journal = {Optimization Methods and Software},
number = {5},
pages = {535--541},
volume = {18},
year = {2003}
}

PowerModelsDistribution.relaxation_psd_to_soc_real_conic — Method

relaxation_psd_to_soc_real_conic(m, mx)

SDP to SOC relaxation of type 2, applied to real-value matrix, as described in:

@article{Kim2003,
author = {Kim, S and Kojima, M and Yamashita, M},
title = {{Second order cone programming relaxation of a positive semidefinite constraint}},
doi = {10.1080/1055678031000148696},
journal = {Optimization Methods and Software},
number = {5},
pages = {535--541},
volume = {18},
year = {2003}
}

Miscellaneous Helpers

PowerModelsDistribution.calculate_tm_scale — Function

Calculates the tap scale factor for the non-dimensionalized equations.

PowerModelsDistribution.calc_branch_y — Function

calc_branch_y(branch::Dict{String,<:Any})

computes branch admittance matrices

PowerModelsDistribution.calc_buspair_parameters — Function

calc_buspair_parameters(buses, branches)

Computes the buspair parameters for the network